% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmb-methods.R
\name{AD}
\alias{AD}
\alias{\%*\%,ad,matrix-method}
\alias{matmul}
\alias{max.advector}
\alias{max}
\alias{min.advector}
\alias{min}
\title{Additional methods for AD types}
\usage{
\S4method{\%*\%}{ad,matrix}(x, y)

\method{max}{advector}(..., na.rm)

\method{min}{advector}(..., na.rm)
}
\arguments{
\item{x}{AD object}

\item{y}{Non-AD matrix}

\item{...}{Objects of class advector}

\item{na.rm}{Not used}
}
\description{
Methods for RTMB AD class
}
\section{Functions}{
\itemize{
\item \code{x \%*\% y}: Matrix product function implemented for mixed AD and non-AD objects with \code{colSums(x * y)}. See \link[RTMB]{ADmatrix}.

\item \code{max(advector)}: Finds the maximum value using \code{\link[=CondExpGt]{CondExpGt()}} over a loop

\item \code{min(advector)}: Finds the minimum value using \code{\link[=CondExpLt]{CondExpLt()}} over a loop

}}
\keyword{internal}
