% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airbnb}
\alias{airbnb}
\title{Airbnb Dataset}
\format{
A data frame with 337 rows and 9 variables:
\describe{
\item{latitude}{latitude coordinate}
\item{longitude}{longitude coordinate}
\item{room_type}{type of room listed}
\item{price}{nightly price in USD}
\item{minimum_nights}{minimum nights required}
\item{number_of_reviews}{number of reviews}
\item{reviews_per_month}{average reviews per month}
\item{calculated_host_listings_count}{number of listings by host}
\item{availability_365}{days available per year}
}
}
\source{
\url{https://www.kaggle.com/datasets/dgomonov/new-york-city-airbnb-open-data/}
}
\usage{
airbnb
}
\description{
A dataset of 337 Airbnb listings in New York City, with 9 variables including latitude, longitude, room type, price, minimum nights, number of reviews, reviews per month, calculated host listings count, and availability in the next 365 days.
}
\keyword{datasets}
