% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albedo.R
\name{albedo}
\alias{albedo}
\title{The albedo function.}
\usage{
albedo(latitude, longitude, tau, force_multiple_ten = FALSE)
}
\arguments{
\item{latitude}{Planetary latitude, from -90° to io +90°.}

\item{longitude}{Planetary longitude, from -180° to 180°.}

\item{tau}{Optical depth.}

\item{force_multiple_ten}{}
}
\value{

}
\description{
Calculate the albedo value given geographical location and tau factor.
}
\details{
Source: Table I - The albedo function.
  Appelbaum, Joseph & Landis, Geoffrey & Sherman, I. (1991).
  Solar radiation on Mars — Update 1991. Solar Energy. 50. 35-51. 10.1016/0038-092X(93)90006-A:
  https://www.researchgate.net/publication/223850868_Solar_radiation_on_Mars-Update_1991
}
