% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.mars}
\alias{anova.mars}
\title{Anova method for MARS}
\usage{
\method{anova}{mars}(object, ...)
}
\arguments{
\item{object}{an object of class 'mars'.}

\item{...}{other arguments -- currently not used.}
}
\description{
Compute an analysis of variance table for the fitted MARS model.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
anova.mars(fit.mars)

}
