% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{bwd_stepwise}
\alias{bwd_stepwise}
\title{Backward stepwise}
\usage{
bwd_stepwise(fwd, control)
}
\arguments{
\item{fwd}{from fwd_stepwise}

\item{control}{mars.control object}
}
\value{
list with y, B and Bfuncs
}
\description{
Backward stepwise
}
