% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_int.R
\name{calc_F}
\alias{calc_F}
\title{Newton-Raphson search for fishing mortality}
\usage{
calc_F(
  Cobs,
  N,
  sel,
  wt,
  M,
  fleet_area,
  q_fs,
  delta = 1,
  na = dim(N)[1],
  nr = dim(N)[2],
  ns = dim(N)[3],
  nf = length(Cobs),
  Fmax = 2,
  nitF = 5L,
  trans = c("log", "logit")
)
}
\arguments{
\item{Cobs}{Observed catch. Vector by \code{f}}

\item{N}{Stock abundance at the beginning of the time step. Array \verb{[a, r, s]}}

\item{sel}{Selectivity. Array \verb{[a, f, s]}}

\item{wt}{Fishery weight at age. Array \verb{[a, f, s]}}

\item{M}{Instantaneous natural mortality. Units of per year \verb{[a, s]}}

\item{fleet_area}{Integers that specify the region \code{r} in which fleet \code{f} operates. Vector by \code{f}}

\item{q_fs}{Relative catchability of stock \code{s} for fleet \code{f}. Matrix \verb{[f, s]}}

\item{delta}{Numeric, the duration of time in years corresponding to the observed catch, e.g., 0.25 is a quarterly time step.}

\item{na}{Integer, number of age classes}

\item{nr}{Integer, number of regions}

\item{ns}{Integer, number of stocks}

\item{nf}{Integer, number of fleets}

\item{Fmax}{Numeric, the maximum Findex value}

\item{nitF}{Integer, number of iterations for the Newton-Raphson routine}

\item{trans}{Whether to perform the search in log or logit space}
}
\value{
A list containing:
\itemize{
\item \code{F_afs} Fishing mortality array
\item \code{F_ars} Fishing mortality array
\item \code{F_index} Index of fishing mortality. Vector by \code{f}
\item \code{CB_fs} Catch (biomass) matrix
\item \code{CN_afs} Catch (abundance) array
\item \code{VB_afs} Vulnerable biomass at the beginning of the time step.
\item \code{penalty} Penalty term returned by \link{posfun} when \code{F_index} exceeds \code{Fmax}
\item \code{fn} Difference between predicted and observed catch at the last iteration. Vector by \code{f}
\item \code{gr} Gradient of \code{fn} with respect to \code{F_index} in either log or logit space at the last iteration. Vector by \code{f}
}
}
\description{
Performs a root finding routine to find the index of F that minimizes the difference between
observed catch and the value predicted by the Baranov equation.
}
\details{
The predicted catch for fleet \code{f} is
\deqn{
C^{\textrm{pred}}_f = \sum_s \sum_a v_{a,f,s} q_{f,s} F_f \dfrac{1 - \exp(-Z_{a,s})}{Z_{a,s}} N_{a,f,s} w_{a,f,s}
}

The Newton-Raphson routine minimizes \eqn{f(\vec{x}) = \vec{C}^{\textrm{pred}} - \vec{C}^{\textrm{obs}}} where the vector arrow indexes fleet.

If \code{trans = "log"}, \eqn{\vec{F} = \exp(\vec{x})}.

If \code{trans = "logit"}, \eqn{\vec{F} = F_{\textrm{max}}/(1 + \exp(-\vec{x}))}.

The gradient with respect to \eqn{\vec{x}} is
\deqn{
f'(\vec{x}) = \sum_s \sum_a v_{a,f,s} q_{f,s} N_{a,f,s} w_{a,f,s} \left(\dfrac{\alpha\gamma}{\beta}\right)'
}

\deqn{
\left(\dfrac{\alpha\gamma}{\beta}\right)' = \dfrac{(\alpha\gamma' + \alpha'\gamma)\beta - \alpha\gamma\beta'}{\beta^2}
}

where

\tabular{l}{
\eqn{\alpha_f = F_f} \cr
\eqn{\beta_{a,s} = Z_{a,s} = M_{a,s} + \sum_f v_{a,f,s} q_{f,s} F_f} \cr
\eqn{\gamma_{a,s} = 1 - \exp(-Z_{a,s})} \cr
\eqn{\beta'_{a,f,s} = v_{a,f} q_{f,s} \alpha'_f} \cr
\eqn{\gamma'_{a,f,s} = \exp(-Z_{a,s})\beta'_{a,f,s}}
}

If \code{trans = "log"}, \eqn{\alpha'_f = \alpha_f}.

If \code{trans = "logit"}, \eqn{\alpha'_f = F_{\textrm{max}}\exp(-x_f)/(1 + \exp(-x_f))^2}.

This function solves for \eqn{\vec{x}} by iterating until \eqn{f(\vec{x})} approaches zero. In iteration \eqn{i+1}:
\deqn{\vec{x}_{i+1} = \vec{x}_i - \dfrac{f(\vec{x}_i)}{f'(\vec{x}_i)}}.
}
\author{
Q. Huynh
}
