% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_int.R
\name{calc_LAK}
\alias{calc_LAK}
\title{Length-at-age key}
\usage{
calc_LAK(len_a, sd_la, lbin, nl = length(lbin) - 1)
}
\arguments{
\item{len_a}{Vector of length-at-age}

\item{sd_la}{Vector of standard deviation in length-at-age}

\item{lbin}{Vector of the lower boundary of the length bins}

\item{nl}{Integer, number of length bins (default is \code{length(lbin) - 1})}
}
\value{
Matrix by age (rows) and length (columns)
}
\description{
Calculates the probability distribution of length-at-age using the normal probability density function
}
