% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fn.R
\name{check_data}
\alias{check_data}
\title{Check dimensions and inputs in MARSdata object}
\usage{
check_data(MARSdata, silent = FALSE)
}
\arguments{
\item{MARSdata}{S4 object containing data inputs. See \linkS4class{MARSdata}}

\item{silent}{Logical, whether or not to report default values to the console}
}
\value{
An updated \linkS4class{MARSdata} object.
}
\description{
Ensures that data inputs are of proper dimension. Whenever possible, default values are added to missing items.
}
\seealso{
\link{MARSdata-class}
}
