% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_solar_time_range.R
\name{constrain_solar_time_range}
\alias{constrain_solar_time_range}
\title{Title}
\usage{
constrain_solar_time_range(
  Ls,
  phi,
  Ts_start,
  Ts_end,
  beta = NULL,
  gamma_c = NULL
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{Ts_start}{}

\item{Ts_end}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
