% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\name{data2}
\alias{data2}
\title{data2}
\description{
The data set contains values for 6 attributes (molecular descriptors) of 908 chemicals used to predict quantitative acute aquatic toxicity towards the fish Pimephales promelas (fathead minnow).
}
\details{
Independent Variables: ClC0 (information indices), SM1_Dz(Z) (2D matrix-based descriptors),
GATS1i (2D autocorrelations), NdsCH (atom-type counts), NdssC (atom-type counts), and
MLOGP (molecule properties).
constraints).
Dependent Variables: LC50 (the concentration that causes death in 50\% of test fish over a 96-hour test duration).
}
\references{
\url{https://archive.ics.uci.edu/ml/datasets/QSAR+fish+toxicity}
}
\keyword{data2}
