% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3.R
\name{data3}
\alias{data3}
\title{data3}
\description{
Concrete is the most important material in civil engineering. Age and ingredients have a highly nonlinear effect on concrete compressive strength.
}
\details{
Independent Variables: Cement (kg in a metric cubic mixture), Blast Furnace Slag (kg in
a metric cubic mixture), Fly Ash (kg in a metric cubic mixture), Water (kg in a metric cubic
mixture), Superplasticizer (kg in a metric cubic mixture), Coarse Aggregate (kg in a metric cubic
mixture), Fine Aggregate (kg in a metric cubic mixture), and Age (days).

Dependent Variable: Concrete compressive strength (MPa).
}
\references{
\url{https://archive.ics.uci.edu/ml/datasets/Concrete+Compressive+Strength}
}
\keyword{data3}
