% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-s4-class.R
\docType{class}
\name{DCKMR-class}
\alias{DCKMR-class}
\title{DCKMR S4 object}
\description{
Store lists of data frames for parent offspring pairs and half-sibling pairs
}
\section{Slots inherited from DCKMR}{

\describe{
\item{\code{POP_s}}{A list by stock of data frames for parent-offspring pairs. Each row in the data frame corresponds to a "sampling unit" defined by the columns:
\tabular{ll}{
\code{a} \tab Capture year of parent \cr
\code{t} \tab Age at capture of parent \cr
\code{y} \tab Birth year of offspring \cr
\code{n} \tab Number of pairwise comparisons \cr
\code{m} \tab Number of POPs \cr
}
}
\item{\code{HSP_s}}{A list by stock of data frames for half-sibling pairs. Each row in the data frame corresponds to a "sampling unit" defined by the columns:
\tabular{ll}{
\code{yi} \tab Birth year of older sibling \cr
\code{yj} \tab Birth year of younger sibling \cr
\code{n} \tab Number of pairwise comparisons \cr
\code{m} \tab Number of HSPs
}
}
\item{\code{CKMR_like}}{Character, likelihood for the POP and HSP sampling units. See \link{like_CKMR} for options.}
}
}

