% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declination.R
\name{declination}
\alias{declination}
\title{Declination angle: the angular position of the Sun at solar noon with respect to the plane of the equator.}
\usage{
declination(Ls, unit = 1)
}
\arguments{
\item{Ls}{Areocentric longitude [deg].}

\item{unit}{}
}
\value{
Declination angle in radians or degrees.
}
\description{
For Mars: -24.936° <= delta <= 24.936°.
0 deg at vernal (Ls 0°) and autumnal equinoxes (Ls 180°)
+24.936 deg at summer solstice (Ls 90) and -24.936 at at winter solstice (Ls 270°).
}
\details{
Source: Equation 7 (1990)
}
