% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Dataset}
\format{
A data frame with 767 rows and 9 variables:
\describe{
\item{X6}{number of pregnancies}
\item{X148}{plasma glucose concentration}
\item{X72}{diastolic blood pressure}
\item{X35}{triceps skin fold thickness}
\item{X0}{2-hour serum insulin}
\item{X33.6}{body mass index}
\item{X0.627}{diabetes pedigree function}
\item{X50}{age in years}
\item{X1}{diabetes diagnosis (0=no, 1=yes)}
}
}
\source{
\url{https://machinelearningmastery.com}
}
\usage{
diabetes
}
\description{
A dataset of 767 patients with diabetes, with 9 variables including number of pregnancies, plasma glucose concentration, diastolic blood pressure, skin thickness, insulin level, body mass index, diabetes pedigree function, age, and diabetes outcome.
}
\keyword{datasets}
