% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-s4-class.R
\docType{class}
\name{Dlabel-class}
\alias{Dlabel-class}
\title{Dlabel S4 object}
\description{
Vectors for labeling plots.
}
\section{Slots inherited from Dlabel}{

\describe{
\item{\code{year}}{Vector of years. Length \code{Dmodel@ny}}
\item{\code{season}}{Vector of season names. Length \code{Dmodel@nm}}
\item{\code{age}}{Vector of ages. Length \code{Dmodel@na}}
\item{\code{region}}{Vector of region names. Length \code{Dmodel@nr}}
\item{\code{stock}}{Vector of stock names. Length \code{Dmodel@ns}}
\item{\code{fleet}}{Vector of fleet names. Length \code{Dfishery@nf}}
\item{\code{index}}{Vector of index of abundance names. Length \code{Dsurvey@ni}}
}
}

