% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-s4-class.R
\docType{class}
\name{Dstock-class}
\alias{Dstock-class}
\title{Dstock S4 object}
\description{
S4 class that organizes the various data inputs for the MARS model. \code{MARSdata} simply inherits the slots from 6 component classes:
\code{Dmodel}, \code{Dstock}, \code{Dfishery}, \code{Dsurvey} \code{DCKMR}, and \code{Dtag}, where the \code{D}- prefix denotes an object for model data.
}
\details{
For convenience, most arrays and matrices have the associated dimensions in the variable name. For example, \code{Cobs_ymfr} represents
observed catch with the dimension following the underscore, following this template:

\tabular{ll}{
\code{y} \tab Year \cr
\code{m} \tab Season \cr
\code{a} \tab Age \cr
\code{r} \tab Region \cr
\code{f} \tab Fishery \cr
\code{i} \tab Index \cr
\code{s} \tab Stock
}
}
\section{Slots inherited from Dstock}{

\describe{
\item{\code{len_ymas}}{Length-at-age. Only needed if \code{Dmodel@nl > 0}.}
\item{\code{sdlen_ymas}}{Standard deviation in length-at-age}
\item{\code{LAK_ymals}}{Length-at-age probability array. If empty, values will be calculated in \link{check_data} (see \link{calc_LAK}).}
\item{\code{mat_yas}}{Proportion mature by age class}
\item{\code{fec_yas}}{Fecundity, i.e., spawning output, of mature animals}
\item{\code{swt_ymas}}{Stock weight-at-age}
\item{\code{Md_yas}}{Natural morality. Ignored if M is estimated.}
\item{\code{m_spawn}}{Integer, season of spawning}
\item{\code{m_rec}}{Integer, season of recruitment}
\item{\code{SRR_s}}{Character vector of stock-recruit relationship by stock. See \link{calc_recruitment} for options.}
\item{\code{delta_s}}{Fraction of season that elapses when spawning occurs}
\item{\code{natal_rs}}{Boolean matrix that indicates whether stock \code{s} spawns when in region \code{r} at time of spawning.}
}
}

\seealso{
\link{MARSdata-class} \code{\link[=check_data]{check_data()}} \link{Dmodel-class} \link{Dstock-class} \link{Dfishery-class} \link{Dsurvey-class} \link{DCKMR-class} \link{Dtag-class}
}
