% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{he analytical expression of the normalized net flux function.}
\usage{
f(z, tau, al = 0.1, pub_year = NULL)
}
\arguments{
\item{z}{Zenith angle [deg].}

\item{tau}{Optical depth.}

\item{al}{Albedo (ranges from 0.1 to 0.4).}

\item{pub_year}{}
}
\value{
Normalized net flux.
}
\description{
Source: Equation 20 (1990).
}
