% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f_89}
\alias{f_89}
\title{The normalized net flux function's lookup table lookup.}
\usage{
f_89(z, tau, al = 0.1)
}
\arguments{
\item{z}{Zenith angle [deg].}

\item{tau}{Optical depth.}

\item{al}{Albedo, can only be 0.1.}
}
\value{
Normalized net flux.
}
\description{
The net solar flux integrated over the solar spectrum on the Martian
surface was calculated by Pollack based on multiple wavelength and
multiple scattering of the solar radiation.
}
\details{
Source: Table III in Appelbaum, Joseph & Flood, Dennis. (1990).
}
