% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f_90}
\alias{f_90}
\title{The normalized net flux function's lookup table lookup.}
\usage{
f_90(z, tau, al = 0.1)
}
\arguments{
\item{z}{Zenith angle [deg].}

\item{tau}{Optical depth.}

\item{al}{Albedo, can be 0.1 or 0.4.}
}
\value{
Normalized net flux.
}
\description{
Source: Solar radiation on Mars: Update 1990.
 Table III (a) and (b) for an albedo of 0.1.
 Table IV  (a) and (b) for an albedo of 0.4.
}
