% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f_analytical}
\alias{f_analytical}
\title{The analytical expression of the normalized net flux function.}
\usage{
f_analytical(z, tau, al = 0.1)
}
\arguments{
\item{z}{Zenith angle [deg].}

\item{tau}{Optical depth.}

\item{al}{Albedo.}
}
\value{
Normalized net flux.
}
\description{
The mean error is about 0.7 percent for the full range. For zenith angles up to 40° the error is much smaller.
The largest error is for zenith angle of 80° and 85° and for tau greater than 5. The maximum error is about 7 percent.
At these large angles and opacities, the error has a minor effect on the calculated daily insolations.
}
\details{
Source: Equation 20? in Appelbaum, Joseph & Flood, Dennis (1990) - Update 1990.
}
