% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{fwd_stepwise}
\alias{fwd_stepwise}
\title{Forward stepwise}
\usage{
fwd_stepwise(y, x, control = mars.control())
}
\arguments{
\item{y}{vector of dependent/response variables}

\item{x}{matrix of independent/predictor variables}

\item{control}{mars.control object}
}
\value{
list with y, B and Bfuncs
}
\description{
Forward stepwise
}
