% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_ali.R
\name{G_ali}
\alias{G_ali}
\title{Title}
\usage{
G_ali(
  Ls,
  phi,
  longitude,
  T_s,
  Z = Z_eq(Ls = Ls, T_s = T_s, phi = phi, nfft = nfft),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta,
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{T_s}{}

\item{Z}{}

\item{tau}{}

\item{al}{}

\item{beta}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
