% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_bh.R
\name{G_bh}
\alias{G_bh}
\title{Title}
\usage{
G_bh(Ls, phi, Ts, z = Z(Ls = Ls, phi = phi, Ts = Ts), tau)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{Ts}{}

\item{z}{}

\item{tau}{}

\item{longitude}{}

\item{al}{}
}
\value{

}
\description{
Title
}
