% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_bi.R
\name{G_bi}
\alias{G_bi}
\title{Title}
\usage{
G_bi(
  Ls,
  phi,
  T_s,
  z = Z(Ls = Ls, T_s = T_s, phi = phi, nfft = nfft),
  tau,
  beta,
  gamma_c,
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}

\item{beta}{}

\item{gamma_c}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
