% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_i.R
\name{G_i}
\alias{G_i}
\title{Title}
\usage{
G_i(
  Ls,
  phi,
  longitude,
  Ts,
  z = Z(Ls = Ls, phi = phi, Ts = Ts),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta,
  gamma_c
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{Ts}{}

\item{z}{}

\item{tau}{}

\item{al}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
