% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_ob.R
\name{G_ob}
\alias{G_ob}
\title{Equation 4 (1990): Beam irridiance at the top of the Martian atmosphere [W/m2].}
\usage{
G_ob(Ls)
}
\arguments{
\item{Ls}{Areocentric longitude.}
}
\value{
Beam irradiance [W/m2].
}
\description{
Source: 
  Appelbaum, Joseph & Flood, Dennis. (1990).
  Solar radiation on Mars. Solar Energy. 45. 353–363. 10.1016/0038-092X(90)90156-7. 
  https://www.researchgate.net/publication/256334925_Solar_radiation_on_Mars
}
