% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_obh.R
\name{G_obh}
\alias{G_obh}
\title{Title}
\usage{
G_obh(Ls, phi = NULL, Ts = NULL, z = Z(Ls = Ls, phi = phi, Ts = Ts))
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{Ts}{}

\item{z}{}
}
\value{

}
\description{
Title
}
