% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_MARSdata}
\alias{get_MARSdata}
\title{Retrieve data object used to fit model}
\usage{
get_MARSdata(MARSassess)
}
\arguments{
\item{MARSassess}{\linkS4class{MARSassess} object returned by \code{fit_MARS()}}
}
\value{
\linkS4class{MARSdata} object
}
\description{
A convenient function to retrieve the data object used to fit the model. The object is embedded in an environment
within the RTMB object.
}
