% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{get_net_flux_function_type}
\alias{get_net_flux_function_type}
\title{Get the net flux function type that is set in the NET_FLUX_FUNCTION_TYPE environment variable.
Can either be 'lookup_v1', 'lookup_v2', or 'polynomial'.
If NET_FLUX_FUNCTION_TYPE has not been set then default to 'polynomial'.}
\usage{
get_net_flux_function_type()
}
\value{

}
\description{
Get the net flux function type that is set in the NET_FLUX_FUNCTION_TYPE environment variable.
Can either be 'lookup_v1', 'lookup_v2', or 'polynomial'.
If NET_FLUX_FUNCTION_TYPE has not been set then default to 'polynomial'.
}
