% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R, R/plot.mars.R, R/predict.mars.R
\name{h}
\alias{h}
\title{Hinge Function}
\usage{
h(x, s, t)

h(x, s, t)

h(x, s, t)
}
\arguments{
\item{x}{vector of predictors}

\item{s}{+/- hinge}

\item{t}{split location}
}
\value{
maximum of inputs

maximum of inputs

maximum of inputs
}
\description{
Hinge Function

Hinge Function

Hinge Function
}
