% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_ali.R
\name{H_ali}
\alias{H_ali}
\title{Title}
\usage{
H_ali(
  Ls,
  phi,
  longitude,
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta,
  gamma_c
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{tau}{}

\item{al}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
