% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_dh.R
\name{H_dh}
\alias{H_dh}
\title{Title}
\usage{
H_dh(
  Ls,
  phi,
  longitude,
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{tau}{}

\item{al}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
