% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_di.R
\name{H_di}
\alias{H_di}
\title{Title}
\usage{
H_di(
  Ls,
  phi,
  longitude,
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta,
  gamma_c,
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{tau}{}

\item{al}{}

\item{beta}{}

\item{gamma_c}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
