% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{heatmap_axis}
\alias{heatmap_axis}
\title{heatmap_axis}
\usage{
heatmap_axis(side = 1, axis_range = range(labels), labels,
  ruler_axis = TRUE, ...)
}
\arguments{
\item{side}{the axis should be plotted on}

\item{axis_range}{numeric range of the axis}

\item{labels}{of each row/column (that is all, not just the ones you want plotted)}

\item{ruler_axis}{do you want to use the \code{\link{ruler_axis}}?}

\item{...}{further parameters for \code{axis(...)}}
}
\description{
Helper function to automatically plot a heatmap axis that makes sense.
}

