% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{housing_prices_test}
\alias{housing_prices_test}
\title{Housing prices in Ames, Iowa - test dataset}
\format{
A data frame of 1147 rows and 24 columns.
}
\usage{
data(housing_prices_test)
}
\description{
This is a dataset with 24 explanatory variables, each describing different aspect of
resedential homes in Ames, Iowa.
}
\details{
\itemize{
\item LotFrontage. Linear feet of street connected to property.
\item LotArea. Lot size in square feet.
\item OverallQual. Rates the overall material and finish of the house. 10 - Very Excellent, 9 - Excellent, 8 - Very Good, 7 - Good, 6 - Above Average, 5 - Average, 4 - Below Average, 3 - Fair, 2 - Poor, 1 - Very Poor.
\item OverallCond. Rates the overall condition of the house. 10 - Very Excellent, 9 - Excellent, 8 - Very Good, 7 - Good, 6 - Above Average, 5 - Average, 4 - Below Average, 3 - Fair, 2 - Poor, 1 - Very Poor.
\item YearBuilt. Original construction date.
\item YearRemodAdd. Remodel date (same as construction date if no remodelling or additions)
\item MasVnrArea. Masonry veneer area in square feet.
\item BsmtUnfSF. Unfinished square feet of basement area.
\item TotalBsmtSF. Total square feet of basement area.
\item X1stFlrSF. First floor square feet.
\item X2ndFlrSF. Second Floor square feet.
\item GrLivArea. Above ground living area square feet.
\item FullBath. Full bathrooms above grade.
\item Halfbath. half baths above grade.
\item BedroomAbvGr. Bedrooms above grade.
\item KitchenAbvGr. Kitchens above grade.
\item TotRmsAbvGrd. Total rooms above grade.
\item Fireplaces. Number of fireplaces.
\item GarageYrBlt. Year garage was built.
\item GarageCars. Size of garage in car capacity.
\item GarageArea. Size of garage in square feet.
\item WoodDeckSF. Wood deck area in square feet.
\item OpenPorchSF. Open porch area in square feet.
\item YrSold. Year Sold (YYYY).
}
}
\references{
https://www.kaggle.com/competitions/house-prices-advanced-regression-techniques/overview
}
\author{
Dean De Cock
}
\keyword{datasets}
