% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_bh.R
\name{I_bh}
\alias{I_bh}
\title{Title}
\usage{
I_bh(Ls, phi, tau, T_start, T_end, al = NULL, nfft)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{tau}{}

\item{T_start}{}

\item{T_end}{}

\item{al}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
