% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_bi.R
\name{I_bi}
\alias{I_bi}
\title{Title}
\usage{
I_bi(Ls, phi, tau, T_start, T_end, beta, gamma_c)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{tau}{}

\item{T_start}{}

\item{T_end}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
