% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_dh.R
\name{I_dh}
\alias{I_dh}
\title{Title}
\usage{
I_dh(
  Ls,
  phi,
  longitude,
  tau,
  T_start,
  T_end,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{tau}{}

\item{T_start}{}

\item{T_end}{}

\item{al}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
