% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_h.R
\name{I_h}
\alias{I_h}
\title{Title}
\usage{
I_h(
  Ls,
  phi,
  longitude,
  tau,
  Ts_start,
  Ts_end,
  al = albedo(latitude = phi, longitude = longitude, tau = tau)
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{tau}{}

\item{Ts_start}{}

\item{Ts_end}{}

\item{al}{}
}
\value{

}
\description{
Title
}
