% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_i.R
\name{I_i}
\alias{I_i}
\title{Title}
\usage{
I_i(
  Ls,
  phi,
  longitude,
  tau,
  T_start,
  T_end,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta,
  gamma_c,
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{tau}{}

\item{T_start}{}

\item{T_end}{}

\item{al}{}

\item{beta}{}

\item{gamma_c}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
