% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{init_B}
\alias{init_B}
\title{Initialize basis functions}
\usage{
init_B(N, Mmax)
}
\arguments{
\item{N}{sample size}

\item{Mmax}{maximum of basis functions}
}
\value{
data frame of N x (Mmax+1)
}
\description{
Initialize basis functions
}
