% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{install_marks_package_suite}
\alias{install_marks_package_suite}
\title{install_marks_package_suite}
\usage{
install_marks_package_suite(auth_user = "markheron")
}
\arguments{
\item{auth_user}{(character) The bitbucket account used to download the packages (default: "marhheron")}
}
\description{
Installs Mark's package suite from bitbucket in one go!
}
