% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{LOF}
\alias{LOF}
\title{Lack of Fit (LOF)}
\usage{
LOF(form, data, control)
}
\arguments{
\item{form}{object of formula}

\item{data}{input data frame}

\item{control}{mars.control object}
}
\value{
value of GCV
}
\description{
Lack of Fit (LOF)
}
