% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{make_B}
\alias{make_B}
\title{New basis functions}
\usage{
make_B(X, Bfuncs)
}
\arguments{
\item{X}{new predictors}

\item{Bfuncs}{basis functions}
}
\value{
new basis functions
}
\description{
New basis functions
}
