% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-parameters.R
\name{make_parameters}
\alias{make_parameters}
\alias{make_map}
\title{Make list of parameters for RTMB}
\usage{
make_parameters(MARSdata, start = list(), ...)

make_map(
  p,
  MARSdata,
  est_M = FALSE,
  est_h = FALSE,
  est_mat = FALSE,
  est_sdr = FALSE,
  est_mov = c("none", "dist_random", "gravity_fixed"),
  est_qfs = FALSE,
  silent = FALSE
)
}
\arguments{
\item{MARSdata}{S4 data object}

\item{start}{An optional list of parameters. Named list of parameters with the associated dimensions and transformations below.
Overrides default values created by \code{\link[=make_parameters]{make_parameters()}}.}

\item{...}{Various arguments for \code{\link[=make_map]{make_map()}} (could be important!)}

\item{p}{List of parameters, e.g., returned by \code{\link[=make_parameters]{make_parameters()}}}

\item{est_M}{Logical, estimate natural mortality?}

\item{est_h}{Logical, estimate steepness?}

\item{est_mat}{Logical, estimate maturity?}

\item{est_sdr}{Logical, estimate standard deviation of recruitment deviates?}

\item{est_mov}{Character describing structure of stock movement parameters. See details below.}

\item{est_qfs}{Logical, estimate relative catchability of stocks by each fleet? Fix \code{log_q_fs} for the first stock if \code{TRUE}}

\item{silent}{Logical, whether to report messages to the console}
}
\value{
\code{\link[=make_map]{make_map()}} returns a named list containing parameter mappings (\code{"map"}) and a character vector of random effects (\code{"random"}).

\code{\link[=make_parameters]{make_parameters()}} returns a list of parameters (\code{"p"}) concatenated with the output of \code{\link[=make_map]{make_map()}}.
}
\description{
Sets up the list of parameters, map of parameters (see \code{map} argument in \code{\link[TMB:MakeADFun]{TMB::MakeADFun()}}), and identifies some random effects parameters
based on the input data and some user choices on model configuration.

These functions provide a template for the parameter and map setup that can be adjusted for alternative configurations. \code{\link[=check_parameters]{check_parameters()}}
checks whether custom made parameter lists are of the correct dimension.
}
\section{Parameters}{


Generally parameter names will have up to three components, separated by underscores.
For example, \code{log_M_s} represents natural mortality in log space is a vector by stock \code{s}.

The first component describes the transformation from the estimated parameter space to the normal parameter space,
for example, \code{log} or \code{logit}. Prefix \code{t} indicates some other transformation that is described below.

Second is the parameter name, e.g., \code{M} for natural mortality, \code{rdev} for recruitment deviates, etc.

Third is the dimension of the parameter variable and the indexing for the vectors, matrices, and arrays, e.g., \code{y} for year, \code{s} for stock.
See \link{MARSdata-class}. Here, an additional index \code{p} represents some other number of parameters that is described below.

\describe{
\item{\code{t_R0_s}}{Vector by \code{s}. Unfished recruitment, i.e., intersection of unfished replacement line and average stock recruit function,
is represented as: \code{R0 <- exp(t_R0_s + MARSdata@Dmodel@scalar_s)}. By default, \code{t_R0_s = 3}}
\item{\code{t_h_s}}{Vector by \code{s}. Steepness of the stock-recruit function. Logit space for Beverton-Holt and log space for Ricker functions.
Default steepness value of 0.8}
\item{\code{mat_ps}}{Matrix \verb{[2, s]}. Maturity parameters (can be estimated or specified in data object). Parameter in the first row represents
the age of 50 percent maturity in logit space: \code{a50_s <- plogis(mat_ps[1, ] * na)}.
In the second row is the age of 95 percent maturity as a logarithmic offset: \code{a95_s <- exp(a50 + mat_ps[2, ])}.
Default \code{a50_s <- 0.5 * na} and \code{a95_s <- a50_s + 1}}
\item{\code{log_M_s}}{Vector by \code{s}. Natural logarithm of natural mortality (can be estimated or specified in data object).
Default parameter value for all stocks: \code{M <- -log(0.05)/MARSdata@Dmodel@na}}
\item{\code{log_rdev_ys}}{Matrix \verb{[y, s]}. Log recruitment deviations. By default, all start values are at zero.}
\item{\code{log_sdr_s}}{Vector by \code{s}. log-Standard deviation of the log recruitment deviations. Default SD = 0.4}
\item{\code{log_q_fs}}{Matrix \verb{[f, s]}. The natural logarithm of \code{q_fs}, the relative fishing efficiency of \code{f} for stock \code{s}.
Equal values imply equal catchability of all stocks. See equations in \code{\link[=calc_F]{calc_F()}}. Default sets all values to zero.}
\item{\code{sel_pf}}{Matrix \verb{[3, f]}. Fishery selectivity parameters in logit or log space. See equations \code{\link[=conv_selpar]{conv_selpar()}}, where \code{sel_pf} is the \code{x} matrix.}
\item{\code{sel_pi}}{Matrix \verb{[3, i]}. Index selectivity parameters in logit or log space. See equations \code{\link[=conv_selpar]{conv_selpar()}}, where \code{sel_pi} is the \code{x} matrix.}
\item{\code{mov_x_ymarrs}}{Array \verb{[y, m, a, r, r, s]}. Base movement matrix. Set to -1000 to effectively exclude movements from region pairs.
See equations in \code{\link[=conv_mov]{conv_mov()}}}
\item{\code{mov_g_ymars}}{Array \verb{[y, m, a, r, s]}. Attractivity term in gravity model for movement. If \code{x} and \code{v} are zero,
this matrix specifies the distribution of total stock abundance into the various regions. See equations in \code{\link[=conv_mov]{conv_mov()}}}
\item{\code{mov_v_ymas}}{Array \verb{[y, m, a, s]}. Viscosity term in gravity model for movement. See equations in \code{\link[=conv_mov]{conv_mov()}}}
\item{\code{log_sdg_rs}}{Array \verb{[r, s]}. Marginal log standard deviation in the stock distribution (\code{mov_g_ymars}) among regions for stock \code{s}.
Only used when \code{est_mov = dist_random}. Default SD of 0.1.}
\item{\code{t_corg_ps}}{Array \verb{[sum(1:(nr - 1)), s]}. Lower triangle of the correlation matrix for \code{mov_g_ymars}, to be obtained with the
Cholesky factorization. Only used when \code{est_mov = dist_random}. Default values of zero.}
}
}

\section{Movement setup for \code{make_map()}}{

If a single region model or \code{est_mov = "none"}: no movement parameters are estimated.

If \code{est_mov = "dist_random"}: fix all values for \code{mov_x_ymarrs} and \code{mov_v_ymas}. Fix \code{mov_g_ymars} for the first region for each year,
season, age, and stock. \code{mov_g_ymars} are random effects.

If \code{est_mov = "gravity_fixed"}: fix all values for \code{mov_x_ymarrs}. Fix \code{mov_g_ymars} for the first region for each year,
season, age, and stock. Estimate all \code{mov_v_ymas}. Both \code{mov_g_ymars} and \code{mov_v_ymas} are fixed effects.
}

\seealso{
\link{MARSdata-class}
}
