% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars-package.R
\docType{package}
\name{mars-package}
\alias{mars-package}
\title{mars: Multivariate Adaptive Regression Splines}
\description{
Implementing Freidman's (1991), Multivariate Adaptive Regression Splines (MARS). Using forward and backward recursive stepwise selection on basis functions for non-linear regression.
}
\author{
\strong{Maintainer}: Kennedy Jokonya \email{zkj@sfu.ca}

}
\keyword{internal}
