% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_extensions.R
\name{Math.ff_matrix}
\alias{Math.ff_matrix}
\title{Math.ff_matrix}
\usage{
\method{Math}{ff_matrix}(x, ...)
}
\arguments{
\item{x}{object}

\item{\dots}{further arguments passed to methods}
}
\description{
Math.ff_matrix
}

