% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmb-methods.R
\name{matmul}
\alias{matmul}
\alias{\%*\%,ad,matrix-method}
\title{Matrix multiplication}
\usage{
\S4method{\%*\%}{ad,matrix}(x, y)
}
\arguments{
\item{x}{AD object}

\item{y}{Matrix}
}
\description{
Matrix product function implemented for mixed AD and non-AD objects. See \link[RTMB]{ADmatrix}.
}
\keyword{internal}
