% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{new_mars.control}
\alias{new_mars.control}
\title{mars.control helper function}
\usage{
new_mars.control(control)
}
\arguments{
\item{control}{a control object}
}
\value{
a mars.control object
}
\description{
Helper function to create a new mars.control object
}
