% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_angle.R
\name{optimal_angle}
\alias{optimal_angle}
\title{Optimal beta angle of the surface for clear and cloudy skies.}
\usage{
optimal_angle(Ls, phi, unit = 1)
}
\arguments{
\item{Ls}{areocentric longitude.}

\item{phi}{planetary latitude.}

\item{unit}{unit of the calculated beta angle, 1 for radians or 2 for degrees.}
}
\value{

}
\description{
Taken from Equations 40 and 43 in Appelbaum 1993.
Warning: 
- Equation 40 in Appelbaum 1993 is wrong. Took Equation 40 in Appelbaum 1995 instead.
- Equation 40 in Appelbaum 1995 returns negative values. Changed the equation to include abs().
}
