% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{p}
\alias{p}
\title{The coefficient lookup function.}
\usage{
p(i, j, k)
}
\arguments{
\item{i}{}

\item{j}{}

\item{k}{}
}
\value{

}
\description{
Source: Table IV - Normalized Net Flux Function Coefficients in Appelbaum, Joseph & Flood, Dennis (1990) Update 1990.
}
