% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{plot.mars}
\usage{
\method{plot}{mars}(
  x,
  n = 20,
  theta = 300,
  phi = 25,
  col = "lightblue",
  border = "black",
  ...
)
}
\arguments{
\item{x}{mars object}

\item{n}{length.out for 3D plot}

\item{theta}{controls horizontal view axis for 3D plot}

\item{phi}{controls vertical view axis for 3D plot}

\item{col}{color of 3D plot}

\item{border}{border color for 3D plot}

\item{...}{additional arguments}
}
\value{
plots of basis functions
}
\description{
plot.mars
}
\examples{
set.seed(2026)
x1 <- 1:100
x2 <- rnorm(100)
x3 <- rnorm(100)
y <- pmax(0, x1 - 0.5)*pmax(0, x2 - 0.5)  + rnorm(100)
data <- data.frame(y,x1,x2,x3)
fit <- mars(y ~ ., data = data, control = mars.control(Mmax=4))
plot(fit)
}
