% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{posfun}
\alias{posfun}
\title{Quadratic penalty function}
\usage{
posfun(x, eps)
}
\arguments{
\item{x}{Numeric, the parameter}

\item{eps}{Numeric, the threshold below which a penalty will be applied}
}
\value{
The penalty value is

\deqn{
\textrm{penalty} =
\begin{cases}
0.1 (x - \varepsilon)^2 & x \le \varepsilon\\
0 & x > \varepsilon
\end{cases}
}
}
\description{
Taped penalty function if \code{x < eps}
}
