% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{predict.mars}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{mars object of model}

\item{newdata}{new data of predictors}
}
\value{
predicted values
}
\description{
predict.mars
}
\examples{
set.seed(2026)
x1 <- 1:100
x2 <- rnorm(100)
x3 <- rnorm(100)
y <- pmax(0, x1 - 0.5)*pmax(0, x2 - 0.5)  + rnorm(100)
data <- data.frame(y,x1,x2,x3)
newdata <- data[,-1] # drop response y
fit <- mars(y ~ ., data = data, control = mars.control(Mmax=4))
predict(fit,newdata)
}
