% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{print.mars objects}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{mars object}

\item{...}{additional arguments}
}
\value{
call and coefficients of mars output
}
\description{
print.mars objects
}
\examples{
set.seed(2026)
x1 <- 1:100
x2 <- rnorm(100)
x3 <- rnorm(100)
y <- pmax(0, x1 - 0.5)*pmax(0, x2 - 0.5)  + rnorm(100)
data <- data.frame(y,x1,x2,x3)
fit <- mars(y ~ ., data = data, control = mars.control(Mmax=4))
print(fit)
}
