% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\alias{prior_h}
\alias{prior_M}
\alias{prior_q}
\title{Priors for MARS model}
\usage{
prior_h(MARSdata, s = 1, m, stdev)

prior_M(MARSdata, s = 1, meanlog, sdlog)

prior_q(MARSdata, i = 1, meanlog, sdlog)
}
\arguments{
\item{MARSdata}{Data object. Class \linkS4class{MARSdata}}

\item{s}{Integer for stock}

\item{m}{Mean in un-transformed space}

\item{stdev}{Standard deviation in un-transformed space}

\item{meanlog}{Mean of the lognormal distribution on the log scale}

\item{sdlog}{Standard of the lognormal distribution on the log scale}

\item{i}{Integer for the corresponding index}
}
\value{
Character.
}
\description{
text
}
\details{
\itemize{
\item \code{prior_h} returns the prior for stock-recruit steepness. Beta distribution for the Beverton-Holt function and normal distribution for Ricker function.
}

\itemize{
\item \code{prior_M} returns the prior for natural mortality. Lognormal distribution.
}

\itemize{
\item \code{prior_q} returns the prior for index catchability. Lognormal distribution.
}
}
\examples{
# Add M and steepness prior to model

dat <- new("MARSdata")
dat@Dmodel@ns <- 1
dat@Dstock@SRR_s <- "BH"

pr_M <- prior_M(dat, s = 1, log(0.25), 0.3)
pr_h <- prior_h(dat, s = 1, 0.8, 0.15)
dat@Dmodel@prior <- c(pr_M, pr_h)
}
