% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-s4-class.R
\name{report}
\alias{report}
\title{Report generic}
\usage{
report(object, ...)
}
\arguments{
\item{object}{An object from MARS.}

\item{...}{Additional arguments to render reports.}
}
\description{
Used to render rmarkdown reports.
}
