% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective.R
\name{retrospective}
\alias{retrospective}
\alias{plot.MARSretro}
\alias{summary.MARSretro}
\alias{report.MARSretro}
\title{Retrospective analysis}
\usage{
retrospective(MARSassess, yret = 1:5)

\method{plot}{MARSretro}(x, var = c("S_yst", "R_yst", "F_yst", "VB_ymft"), s = 1, f = 1, ...)

\method{summary}{MARSretro}(object, by = c("stock", "fleet"), ...)

\method{report}{MARSretro}(
  object,
  filename = "retro",
  dir = tempdir(),
  open_file = TRUE,
  render_args = list(),
  ...
)
}
\arguments{
\item{MARSassess}{\linkS4class{MARSassess} object}

\item{yret}{Vector of integers (greater than zero) specifying the years to remove for the retrospective analysis}

\item{x, object}{Output of \code{retrospective} function}

\item{var}{Character to indicate the metric, the item in the \code{MARSretro} list to be plotted. See details below.}

\item{s}{Integer for the stock index to plot}

\item{f}{Integer for the fleet index to plot}

\item{...}{Not used}

\item{by}{Character indicating whether to calculate to Mohn's rho on stock or fleet-based time series}

\item{filename}{Character string for the name of the markdown and HTML files.}

\item{dir}{The directory in which the markdown and HTML files will be saved.}

\item{open_file}{Logical, whether the HTML document is opened after it is rendered.}

\item{render_args}{List of arguments to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
A \code{MARSretro} object containing a named lists of arrays generated by the retrospective analysis:
\itemize{
\item \code{S_yst} Spawning output array \verb{[y, s, t]} where \code{t} indexes the retrospective peel
\item \code{R_yst} Recruitment array \verb{[y, s, t]}
\item \code{F_yst} Apical fishing mortality \verb{[y, s, t]}
\item \code{VB_ymft} Vulnerable biomass available to each fishery \verb{[y, m, f, t]}
}

\code{plot.MARSretro} returns individual figures using base graphics.

\code{summary.MARSretro} returns a matrix of Mohn's rho.

\code{report.MARSretro} returns a HTML markdown report.
}
\description{
Perform a retrospective analysis, successive removals of most recent years of data to evaluate
consistency in model estimates of biomass, recruitment, etc.

The \code{summary} method returns Mohn's rho and the plot method generates a markdown report.
}
