\name{ruler_axis}
\alias{ruler_axis}
\title{ruler_axis}
\usage{
ruler_axis(side = 1, axis_range_to_zero_one = NULL, data = NULL)
}
\arguments{
  \item{side}{to add the axis to}

  \item{data}{to better estimate the start/end of the axis
  (if not present it uses the plot dimensions)}
}
\description{
Adds an axis that looks like a ruler (with minor ticks
without labels between the large ticks).
}
\author{
Mark Heron
}

