% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmb-methods.R
\name{sapply2}
\alias{sapply2}
\title{\code{sapply2} function}
\usage{
sapply2(X, FUN, ..., USE.NAMES = TRUE)
}
\arguments{
\item{X, FUN, ..., USE.NAMES}{Same arguments as \code{\link[=sapply]{sapply()}}}
}
\description{
An alternate \code{sapply} function with argument \code{simplify = "array"} for convenience.
}
\keyword{internal}
