% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{scale_to_zero_one_range}
\alias{scale_to_zero_one_range}
\title{rescale data}
\usage{
scale_to_zero_one_range(x, x_range = range(x))
}
\arguments{
\item{x}{data to be rescaled}

\item{x_range}{the original range the data can be in, that is mapped to c(0,1)}
}
\description{
the values of x are rescaled so that x_range goes from 0 to 1
}
\author{
Mark Heron
}

