% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smear.R
\name{smear_ff_matrix}
\alias{smear_ff_matrix}
\title{smear_ff_matrix}
\usage{
smear_ff_matrix(x, from = 0, to = 1)
}
\arguments{
\item{x}{ff matrix to smear along the rows}

\item{from}{where to start}

\item{to}{where to go}
}
\description{
smear_ff_matrix
}
\author{
Mark Heron
}
