% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smear.R
\name{smear_ff_simple}
\alias{smear_ff_simple}
\title{smear_ff_simple}
\usage{
smear_ff_simple(x, from = 0, to = 1)
}
\arguments{
\item{x}{ff vector to smear}

\item{from}{where to start}

\item{to}{where to go}
}
\value{
smeared ff vector
}
\description{
Computes a smeared vector, basically a running_sum.
}
\details{
Adds the data shifted onto itself several times, implementation of \code{\link{smear_simple}} for \code{ff_vector}.
}
\author{
Mark Heron
}
