% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smear.R
\name{smear_simple}
\alias{smear_simple}
\title{smear_simple}
\usage{
smear_simple(x, from = 0, to = 1)
}
\arguments{
\item{x}{vector to smear}

\item{from}{where to start}

\item{to}{where to go}
}
\value{
smeared vector
}
\description{
Computes a smeared vector, basically a running_sum.
Simple (and slow) version used to test against.
}
\details{
Adds the vector on its self several times, so good for short smears of long vectors.
}

