% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{split_points}
\alias{split_points}
\title{Split Points}
\usage{
split_points(x, B)
}
\arguments{
\item{x}{vector of predictors}

\item{B}{basis function}
}
\value{
sorted split points on x
}
\description{
Split Points
}
