% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spotify}
\alias{spotify}
\title{Spotify song data}
\format{
A data frame with 350 rows and 15 variables:
\describe{
\item{X}{index}
\item{popularity}{track popularity score 0-100}
\item{duration_ms}{track duration in milliseconds}
\item{danceability}{danceability score 0-1}
\item{energy}{energy level 0-1}
\item{key}{musical key}
\item{loudness}{loudness in dB}
\item{mode}{major or minor mode}
\item{speechiness}{speechiness score 0-1}
\item{acousticness}{acousticness score 0-1}
\item{instrumentalness}{instrumentalness score 0-1}
\item{liveness}{liveness score 0-1}
\item{valence}{musical positivity score 0-1}
\item{tempo}{beats per minute}
\item{time_signature}{time signature}
}
}
\source{
\url{https://www.kaggle.com/datasets/maharshipandya/spotify-tracks-dataset}
}
\usage{
spotify
}
\description{
A dataset of 350 songs with 15 variables, including popularity, duration, danceability, energy, key, loudness, mode, speechiness, acousticness, instrumentalness, liveness, valence, tempo, and time signature.
}
\keyword{datasets}
