% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_extensions.R
\name{Summary.ff_matrix}
\alias{Summary.ff_matrix}
\title{Summary.ff_matrix}
\usage{
\method{Summary}{ff_matrix}(e1, ...)
}
\arguments{
\item{e1}{object}

\item{\dots}{further arguments passed to methods}
}
\description{
Summary.ff_matrix
}

