% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{summary.mars}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{mars object}

\item{...}{additional arguments}
}
\value{
summary of mars output
}
\description{
summary.mars
}
\examples{
set.seed(2026)
x1 <- 1:100
x2 <- rnorm(100)
x3 <- rnorm(100)
y <- pmax(0, x1 - 0.5)*pmax(0, x2 - 0.5)  + rnorm(100)
data <- data.frame(y,x1,x2,x3)
fit <- mars(y ~ ., data = data, control = mars.control(Mmax=4))
summary(fit)
}
