% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise.R
\name{sunrise}
\alias{sunrise}
\title{Title}
\usage{
sunrise(Ls, phi, beta = NULL, gamma_c = NULL, unit = 1)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{beta}{}

\item{gamma_c}{}

\item{unit}{}
}
\value{

}
\description{
Title
}
