% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise.R
\name{sunrise_for_horizontal_surface}
\alias{sunrise_for_horizontal_surface}
\title{Title}
\usage{
sunrise_for_horizontal_surface(phi, delta)
}
\arguments{
\item{phi}{}

\item{delta}{}
}
\value{

}
\description{
Title
}
