% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise.R
\name{sunrise_for_inclined_surface_oriented_equator}
\alias{sunrise_for_inclined_surface_oriented_equator}
\title{Title}
\usage{
sunrise_for_inclined_surface_oriented_equator(phi, beta, delta)
}
\arguments{
\item{phi}{}

\item{beta}{}

\item{delta}{}
}
\value{

}
\description{
Title
}
