% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunrise.R
\name{sunrise_for_inclined_surface_oriented_west}
\alias{sunrise_for_inclined_surface_oriented_west}
\title{Title}
\usage{
sunrise_for_inclined_surface_oriented_west(phi, beta, gamma_c, delta)
}
\arguments{
\item{phi}{}

\item{beta}{}

\item{gamma_c}{}

\item{delta}{}
}
\value{

}
\description{
Title
}
