% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset_for_horizontal_surface}
\alias{sunset_for_horizontal_surface}
\title{Get the moment in which the sunset occurs.}
\usage{
sunset_for_horizontal_surface(phi, delta)
}
\arguments{
\item{phi}{}

\item{delta}{}
}
\value{

}
\description{
Get the moment in which the sunset occurs.
}
