% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset_for_inclined_surface}
\alias{sunset_for_inclined_surface}
\title{Title}
\usage{
sunset_for_inclined_surface(phi, beta, gamma_c, delta)
}
\arguments{
\item{phi}{}

\item{beta}{}

\item{gamma_c}{}

\item{delta}{}
}
\value{

}
\description{
Title
}
