% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset.R
\name{sunset_for_inclined_surface_oriented_equator}
\alias{sunset_for_inclined_surface_oriented_equator}
\title{(31) in (1993).}
\usage{
sunset_for_inclined_surface_oriented_equator(phi, beta, delta)
}
\arguments{
\item{phi}{}

\item{beta}{}

\item{delta}{}
}
\value{

}
\description{
(31) in (1993).
}
