% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testmars}
\alias{testmars}
\title{Test object of class "mars" to test summary.mars, print.mars, and plot.mars}
\format{
A list of 18 elements.
}
\usage{
data(testmars)
}
\description{
A list of length 18 which is an object of class "mars". This particular object is created to
run the examples written for methods for an object of class "mars". These methods are" summary.mars, plot.mars,
print.mars.
}
\author{
Dr. Brad McNeney \email{brad_mcneney@sfu.ca}
}
\keyword{datasets}
