% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testmc}
\alias{testmc}
\title{Test object of class "mars.control" to test "mars" function}
\format{
A list with three elements. One is a double, one is an integer and one is logical value.
}
\usage{
data(testmc)
}
\description{
A list of length 3, which is an object of class "mars.control", entered in control
argument of mars function. This particular mars.control object is used in examples
shown in documentation of mars function.
}
\details{
\itemize{
\item Mmax. An integer value representing maximum number of basis functions that a user can accept. In this example object it has a value of 10.
\item d. A double value representing an important component used in Generalized Cross Validation (GCV) criterion of mars function. In this example, the value is 3.
\item trace. A logical value entered by user. TRUE if they want to see what is happening as the mars function is progressing, FALSE if not. In this example, its FALSE.
}
}
\author{
Dr. Brad McNeney \email{brad_mcneney@sfu.ca}
}
\keyword{datasets}
