% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z.R
\name{Z}
\alias{Z}
\title{Equation 6: Zenith angle of the incident solar radiation [deg].}
\usage{
Z(Ls, phi, Ts)
}
\arguments{
\item{Ls}{Areocentric longitude [deg].}

\item{phi}{Planetary latitude [deg].}

\item{Ts}{Solar time [h]}
}
\value{

}
\description{
Source: Equation 6 in Appelbaum, Joseph & Flood, Dennis. (1990).
}
