% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.mars}
\alias{anova.mars}
\title{Anova for MARS}
\usage{
\method{anova}{mars}(object, ...)
}
\arguments{
\item{object}{An object of class mars.}

\item{...}{Other arguments.}
}
\value{
The anova table of the fitted MARS model.
}
\description{
Anova for MARS
}
