% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.mars}
\alias{anova.mars}
\alias{anova}
\title{Anova for MARS}
\usage{
\method{anova}{mars}(object, ...)
}
\arguments{
\item{object}{An object of class mars.}

\item{...}{Other arguments.}
}
\value{
The anova table of the fitted MARS model.
}
\description{
Compute an analysis of variance table for the fitted MARS model.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
anova.mars(fit.mars)
}
\seealso{
\code{\link[=mars]{mars()}}

\code{\link[=print.mars]{print.mars()}}

\code{\link[=summary.mars]{summary.mars()}}

\code{\link[=predict.mars]{predict.mars()}}

\code{\link[=plot.mars]{plot.mars()}}
}
