% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{anova.mars}
\alias{anova.mars}
\title{Conducts analysis of variance on a mars object}
\usage{
\method{anova}{mars}(object)
}
\arguments{
\item{object}{an mars objects}
}
\value{
A data frame containing the sum of squares, F-value, p-value, etc.
}
\description{
Conducts analysis of variance on a mars object
}
\examples{
# see concrete strength ex. in testfile
 anova.mars(object)
}
