% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{anova.mars}
\alias{anova.mars}
\title{Conducts analysis of variance on a mars object}
\usage{
\method{anova}{mars}(object)
}
\arguments{
\item{object}{an mars objects}
}
\value{
A data frame containing the sum of squares, F-value, p-value, etc.
}
\description{
Conducts analysis of variance on a mars object
}
\examples{
# concrete strength
df2 <- read.csv("./Concrete_Data.csv",header=TRUE)
names(df2) <- c("Cement","Blast.Furnace.Slag","Fly.Ash","Water",
                "Superplasticizer","Coarse.Aggregate","Fine.Aggregate.",
                "Age.(day)", "Concrete.compressive.strength")
df2 <- head(df2,500)
object <- mars(Concrete.compressive.strength~.,df2)
mars.anova(object)
}
