% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.mars.R
\name{anova.mars}
\alias{anova.mars}
\title{ANVOA function}
\usage{
\method{anova}{mars}(object)
}
\arguments{
\item{object}{an \code{mars} object from the result from mars() function}
}
\value{
returns an object of class \code{anova}. These objects represent analysis-of-variance and analysis-of-deviance tables.
}
\description{
Compute analysis of variance (or deviance) tables for fitted \code{mars} model objects.
}
\examples{
test <- mars(y~.,data=data1[1:100,])

anova(test)
}
