% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{bwd_stepwise}
\alias{bwd_stepwise}
\title{helper for mars function
does the backward step for mars}
\usage{
bwd_stepwise(fwd, control)
}
\arguments{
\item{fwd}{the results for the fwd stepwise}

\item{controls}{a mars control object}
}
\value{
list with y values, B values and Bfuncs
}
\description{
helper for mars function
does the backward step for mars
}
