% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_int.R
\name{calc_nextN}
\alias{calc_nextN}
\title{Project stock abundance to the next time step}
\usage{
calc_nextN(
  N,
  surv,
  na = dim(N)[1],
  nr = dim(N)[2],
  ns = dim(N)[3],
  advance_age = TRUE,
  R = numeric(ns),
  type = c("dist", "mov"),
  dist = array(1/nr, c(na, nr, ns)),
  mov
)
}
\arguments{
\item{N}{Abundance at current time step. Array \verb{[a, r, s]}}

\item{surv}{Survival during the current time step. Array \verb{[a, r, s]}}

\item{na}{Integer, number of age classes}

\item{nr}{Integer, number of regions}

\item{ns}{Integer, number of stocks}

\item{advance_age}{Logical, whether the animals advance to their next age class}

\item{R}{Incoming recruitment. Only assigned if \code{advance_age = TRUE}. Vector of \code{s}}

\item{type}{Character that indicates whether the stock distribution is assigned based on a distribution array or a movement array.}

\item{dist}{Distribution of the stock in the next time step. Array \verb{[a, r, s]}}

\item{mov}{Movement array in the next time step. Array \verb{[a, r, r, s]}}
}
\value{
Abundance at the next time step. Array \verb{[a, r, s]}
}
\description{
This function generates the abundance array by calculating survival from current mortality, then advances
age classes, re-distributes the stock, and adds recruitment.
}
