% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_int.R
\name{conv_mov}
\alias{conv_mov}
\title{Calculate movement matrix for all age classes}
\usage{
conv_mov(x, g, v, nr = dim(x)[2], na = dim(x)[1], aref = ceiling(0.5 * na))
}
\arguments{
\item{x}{Base log-movement parameters. See details. Array \verb{[na, nr, nr]}}

\item{g}{Gravity model attractivity term. Vector by \code{a}}

\item{v}{Gravity model viscosity term. Vector by \code{a}}

\item{nr}{Integer, number of regions}

\item{na}{Integer, number of ages}

\item{aref}{Integer, reference age class}
}
\value{
Movement array \verb{[a, r, r]}
}
\description{
Movement matrices are calculated for all age classes from a base matrix and a gravity model formulation
(Carruthers et al. 2016).
}
\details{
Rows index region of origin and columns denote region of destination.

In log space, the movement matrix \eqn{m} for age class \eqn{a} from region \eqn{r} to \eqn{r'} is the sum of base matrix \eqn{x} and
gravity matrix \eqn{G}:
\deqn{m_{r,r'} = x_{r,r'} + G_{r,r'}}

To essentially exclude movement from \eqn{r} to \eqn{r'}, set \eqn{x_{r,r'} = -1000}

Gravity matrix \eqn{G} includes an attractivity term \eqn{g} and viscosity term \eqn{v}:

\deqn{G_{r,r'} =
\begin{cases}
g'_a + v_a \quad & r = r'\\
g'_a \quad & \textrm{otherwise}
\end{cases}
}

Vector \eqn{g'} are offset terms relative to the value for the reference age class:
\deqn{g'_a =
\begin{cases}
g_a \quad & a = a_{ref}\\
g_a + g_{aref} \quad & \textrm{otherwise}
\end{cases}
}

The movement matrix in normal space is obtained by the softmax transformation
\deqn{M_{r,r'} = \dfrac{\exp(m_{r,r'})}{\sum_r\exp(m_{r,r'})}}
}
\references{
Carruthers, T.R., et al. 2015. Modelling age-dependent movement: an application to red and
gag groupers in the Gulf of Mexico. CJFAS 72: 1159-1176. \doi{10.1139/cjfas-2014-0471}
}
