% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\name{data2}
\alias{data2}
\title{data2}
\description{
In the Delft data set, 308 full-scale experiments were performed at the Delft Ship Hydromechanics Laboratory. The experiments include 22 different hull forms that are derived from a parent form closely related to Frans Maas' Standfast 43'.
}
\details{
Independent Variables: Longitudinal position of the center of buoyancy, Prismatic coefficient, Length-displacement ratio, Beam-draught ratio, Length-beam ratio , and Froude number.
Dependent Variables: Residuary resistance per unit weight of displacement.
}
\references{
\url{https://archive.ics.uci.edu/ml/datasets/Yacht+Hydrodynamics}
}
\keyword{data2}
