% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{demo.pal}
\alias{demo.pal}
\title{Demo plot for palette}
\usage{
demo.pal(colours, border = (if (n < 32) "light gray" else NA),
  main = paste("color palettes;  n=", n))
}
\arguments{
\item{colours}{color palett}

\item{border}{color of the boarder around the rectangles}

\item{main}{figure title}
}
\description{
Simply plots rectangles with the palette colours.
Method idea and initial code from \emph{LSD} package.
}
\examples{
demo.pal(colours=distinctive_colors(32))

}
\seealso{
distinctive_colors
}
