% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-s4-class.R
\docType{class}
\name{Dmodel-class}
\alias{Dmodel-class}
\title{Dmodel S4 object}
\description{
S4 class that organizes the various data inputs for the MARS model. \code{MARSdata} simply inherits the slots from 6 component classes:
\code{Dmodel}, \code{Dstock}, \code{Dfishery}, \code{Dsurvey} \code{DCKMR}, and \code{Dtag}, where the \code{D}- prefix denotes an object for model data.
}
\details{
For convenience, most arrays and matrices have the associated dimensions in the variable name. For example, \code{Cobs_ymfr} represents
observed catch with the dimension following the underscore, following this template:

\tabular{ll}{
\code{y} \tab Year \cr
\code{m} \tab Season \cr
\code{a} \tab Age \cr
\code{r} \tab Region \cr
\code{f} \tab Fishery \cr
\code{i} \tab Index \cr
\code{s} \tab Stock
}
}
\section{Slots inherited from Dmodel}{

\describe{
\item{\code{ny}}{Integer, number of years}
\item{\code{nm}}{Integer, number of seasons}
\item{\code{na}}{Integer, number of ages}
\item{\code{nl}}{Integer, number of length bins. Set to zero if lengths are not modeled.}
\item{\code{nr}}{Integer, number of spatial regions}
\item{\code{ns}}{Integer, number of stocks}
\item{\code{lbin}}{Vector of lower boundary of length bins. Length \code{nl + 1}}
\item{\code{lmid}}{Vector of midpoint of length bins. Length \code{nl}}
\item{\code{Fmax}}{Numeric, maximum allowable instantaneous fishing mortality rate (units of per season). Defaults to 3.}
\item{\code{nitF}}{Integer, number of iterations to solve Baranov catch equation from observed catch. Defaults to 5.}
\item{\code{dist_type}}{Character. Whether to estimate seasonal stock distribution with a movement matrix \code{"mov"} or re-distributing total abundance by a distribution vector \code{"dist"}}
\item{\code{y_phi}}{Integer, the year from which to obtain values of natural mortality and fecundity for the unfished stock-recruit replacement line (\code{phi}). Relevant if natural mortality or fecundity are time-varying. Defaults to 1.}
}
}

\seealso{
\link{MARSdata-class} \code{\link[=check_data]{check_data()}} \link{Dmodel-class} \link{Dstock-class} \link{Dfishery-class} \link{Dsurvey-class} \link{DCKMR-class} \link{Dtag-class}
}
