% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars_301279712.R
\name{fwd_stepwise}
\alias{fwd_stepwise}
\title{MARS  forward stepwise function}
\usage{
fwd_stepwise()
}
\arguments{
\item{formula}{a R formula}

\item{data}{data to fit the model to}
}
\value{
model
}
\description{
MARS  forward stepwise function
}
\examples{
 mm <- mars (wage~age,data=ISLR::Wage)

}
