% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{fwd_stepwise}
\alias{fwd_stepwise}
\title{Helper function for mars function}
\usage{
fwd_stepwise(y, x, control = mars.control())
}
\arguments{
\item{y}{a vector of y values}

\item{x}{data frame with the x values}
}
\value{
returns a list with y values, B, and Bfuncs
}
\description{
Helper function for mars function
}
