% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_ali.R
\name{G_ali}
\alias{G_ali}
\title{Title}
\usage{
G_ali(
  Ls,
  phi,
  longitude,
  Ts,
  z = Z(Ls = Ls, Ts = Ts, phi = phi),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{Ts}{}

\item{z}{}

\item{tau}{}

\item{al}{}

\item{beta}{}
}
\value{

}
\description{
Title
}
