% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_b.R
\name{G_b}
\alias{G_b}
\title{Title}
\usage{
G_b(Ls, phi = NULL, T_s = NULL, z = Z(Ls = Ls, phi = phi, T_s = T_s), tau)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}
}
\value{

}
\description{
Title
}
