% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_b.R
\name{G_b}
\alias{G_b}
\title{Title}
\usage{
G_b(Ls, phi = NULL, Ts = NULL, z = Z(Ls = Ls, phi = phi, Ts = Ts), tau)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{Ts}{}

\item{z}{}

\item{tau}{}
}
\value{

}
\description{
Title
}
