% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_bh.R
\name{G_bh}
\alias{G_bh}
\title{Title}
\usage{
G_bh(
  Ls,
  phi,
  longitude = NULL,
  T_s,
  z = Z(Ls, T_s, phi, nfft),
  tau,
  al = NULL,
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}

\item{al}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
