% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_bh.R
\name{G_bh}
\alias{G_bh}
\title{Title}
\usage{
G_bh(Ls, phi, T_s, z = Z(Ls = Ls, phi = phi, T_s = T_s), tau)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}

\item{longitude}{}

\item{al}{}
}
\value{

}
\description{
Title
}
