% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_bi.R
\name{G_bi}
\alias{G_bi}
\title{Title}
\usage{
G_bi(Ls, phi, Ts, z = Z(Ls = Ls, phi = phi, Ts = Ts), tau, beta, gamma_c)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{Ts}{}

\item{z}{}

\item{tau}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
