% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_dh.R
\name{G_dh}
\alias{G_dh}
\title{Title}
\usage{
G_dh(
  Ls,
  phi,
  longitude,
  T_s = NULL,
  z = Z(Ls, T_s, phi, nfft),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  nfft
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}

\item{al}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
