% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_dh.R
\name{G_dh}
\alias{G_dh}
\title{Title}
\usage{
G_dh(
  Ls,
  phi,
  longitude,
  Ts = NULL,
  z = Z(Ls = Ls, phi = phi, Ts = Ts),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau)
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{Ts}{}

\item{z}{}

\item{tau}{}

\item{al}{}
}
\value{

}
\description{
Title
}
