% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_di.R
\name{G_di}
\alias{G_di}
\title{Title}
\usage{
G_di(
  Ls,
  phi,
  longitude,
  T_s,
  z = Z(Ls = Ls, phi = phi, T_s = T_s),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}

\item{al}{}

\item{beta}{}
}
\value{

}
\description{
Title
}
