% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_h.R
\name{G_h}
\alias{G_h}
\title{Title}
\usage{
G_h(
  Ls,
  phi,
  longitude,
  T_s = NULL,
  z = Z(Ls = Ls, phi = phi, T_s = T_s),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau)
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}

\item{al}{}
}
\value{

}
\description{
Title
}
