% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_i.R
\name{G_i}
\alias{G_i}
\title{Title}
\usage{
G_i(
  Ls,
  phi,
  longitude,
  T_s,
  z = Z(Ls = Ls, phi = phi, T_s = T_s),
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta,
  gamma_c
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{T_s}{}

\item{z}{}

\item{tau}{}

\item{al}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
