% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G_obh.R
\name{G_obh}
\alias{G_obh}
\title{Title}
\usage{
G_obh(Ls, phi = NULL, T_s = NULL, z = Z(Ls, T_s, phi, nfft), nfft = 1)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{T_s}{}

\item{z}{}

\item{nfft}{}
}
\value{

}
\description{
Title
}
