% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_h.R
\name{H_h}
\alias{H_h}
\title{Title}
\usage{
H_h(
  Ls,
  phi,
  longitude,
  tau,
  al = albedo(latitude = phi, longitude = longitude, tau = tau)
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{tau}{}

\item{al}{}
}
\value{

}
\description{
Title
}
