% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I_di.R
\name{I_di}
\alias{I_di}
\title{Title}
\usage{
I_di(
  Ls,
  phi,
  longitude,
  tau,
  Ts_start,
  Ts_end,
  al = albedo(latitude = phi, longitude = longitude, tau = tau),
  beta,
  gamma_c
)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{longitude}{}

\item{tau}{}

\item{Ts_start}{}

\item{Ts_end}{}

\item{al}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
