% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{init_B}
\alias{init_B}
\title{Helper function for fwd stepwise
initializes the b matrix size N,Mmax}
\usage{
init_B(N, Mmax)
}
\arguments{
\item{N}{number of rows}

\item{Mmax}{number of columns}
}
\value{
a NxMmax matrix
}
\description{
Helper function for fwd stepwise
initializes the b matrix size N,Mmax
}
