% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/other_utils.R
\name{install_marks_package_suite}
\alias{install_marks_package_suite}
\title{install_marks_package_suite}
\usage{
install_marks_package_suite()
}
\description{
installs marks package suite from bitbucket in one go!
}

