% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_irradiated.R
\name{is_irradiated}
\alias{is_irradiated}
\title{Check if there is solar irradiance at the given location and moment.}
\usage{
is_irradiated(Ls, phi, T_s, z = Z(Ls, T_s, phi), beta = NULL, gamma_c = NULL)
}
\arguments{
\item{Ls}{}

\item{phi}{}

\item{T_s}{}

\item{z}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Check if there is solar irradiance at the given location and moment.
}
