% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like.R
\name{like_comp}
\alias{like_comp}
\title{Likelihood for length or age composition vectors}
\usage{
like_comp(
  obs,
  pred,
  N = sum(obs),
  type = c("multinomial", "dirmult1", "dirmult2"),
  ...
)
}
\arguments{
\item{obs}{A vector of observed values}

\item{pred}{A vector of predicted values. Same length as \code{obs}}

\item{N}{Numeric, the sample size corresponding to \code{obs}}

\item{type}{Character for the desired distribution}

\item{...}{Other arguments depending on \code{type}. See below.}
}
\value{
Numeric representing the log-likelihood.
}
\description{
Returns the log-likelihood for composition data, with various statistical distributions supported.
}
\details{
For \code{type = "dirmult1"} or \code{"dirmult2"}, provide \code{p} which represents the linear or saturating parameter, respectively.
}
\examples{
M <- 0.1
age <- seq(1:10)
pred <- exp(-M * age)
obs <- pred * rlnorm(10, sd = 0.05)
like_comp(obs, pred, N = 10, type = "multinomial", N = 10)
like_comp(obs, pred, N = 100, type = "multinomial", N = 100)
like_comp(obs, pred, N = 10, type = "dirmult1", p = 1)
like_comp(obs, pred, N = 10, type = "dirmult1", p = 20)
}
\references{
Thorson et al. 2017. Model-based estimates of effective sample size in stock assessment models using the
Dirichlet-multinomial distribution. Fish. Res. 192:84-93. \doi{10.1016/j.fishres.2016.06.005}
}
