% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{LOF}
\alias{LOF}
\title{helper for fwd_stepwise and bwd_stepwise
comptues the LOF for a form and conrol}
\usage{
LOF(form, data, control)
}
\arguments{
\item{form}{formula used}

\item{data}{a data frame}

\item{control}{a mars control object}
}
\value{
number for error
}
\description{
helper for fwd_stepwise and bwd_stepwise
comptues the LOF for a form and conrol
}
