% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARS-package.R
\docType{package}
\name{MARS-package}
\alias{MARS}
\alias{MARS-package}
\title{MARS: Multi-stock Assessment with Regional Spatiotemporal dynamics}
\description{
Implementation of a next-generation, multi-stock assessment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://blue-matter.github.io/MARS/}
  \item \url{https://github.com/Blue-Matter/MARS}
  \item Report bugs at \url{https://github.com/Blue-Matter/MARS/issues}
}

}
\author{
\strong{Maintainer}: Quang Huynh \email{quang@bluematterscience.com} (\href{https://orcid.org/0000-0001-7835-4376}{ORCID})

}
\keyword{internal}
