% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{Constructor for `mars.control` objects}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{Maximum number of basis functions. Should be an even integer. Default value is 2.}

\item{d}{A smoothing parameter}

\item{trace}{logical, if TRUE traceback??????}
}
\description{
This function constructs a `mars.control` object that specifies
parameters used in the model fitting procedure.
}
