% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{mars.control}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{an integer representing maximum number of basis functions allowed by the user, default value is 2}

\item{d}{smoothing parameter}

\item{trace}{logical value. if user wants to see what is happening in the mars function, then set it to TRUE. Default value is FALSE}
}
\value{
an empty list
}
\description{
mars.control
}
\examples{
# example 1
mc <- mars.control(Mmax=10)
mc

# example 2
# when Mmax is not even
try(mars.control(Mmax=3))

# example 3
# when Mmax is not even, it will round up
try(mars.control(Mmax=5))

}
