% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{Creating objects of class "mars.control"}
\usage{
mars.control(Mmax, d, trace)
}
\arguments{
\item{Mmax}{an integer representing maximum number of basis functions allowed by the user, default value is 2}

\item{d}{smoothing parameter}

\item{trace}{logical value. if user wants to see what is happening in the mars function, then set it to TRUE. Default value is FALSE}
}
\value{
an object of class "mars.control". a list containing three elements
}
\description{
This function is used to create an object of class "mars.control".
}
\details{
this function is used to create an object of class "mars.control", which is later used in the control argument of function "mars".
This function perform certain checks to make sure that Mmax is an even integer. If it is not integer, it stops. If it is not even, it adds one
to the supplied value to make it even. It also stops if d is not a number and trace is not logical.
}
\examples{
mars.control(Mmax = 10, d = 3, trace = FALSE)
}
\references{
Friedman, J.H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, Vol. 19, No. 1 (Mar., 1991), pp. 1-67
}
\seealso{
Multivariate Adaptive Regression Splines (MARS)
}
\author{
Samir Arora and Gurpal Singh Tulli
}
