% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{Constructor for \code{mars.control} objects}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{Maximum number of basis functions for the forward. Should be an even integer. Default value is 2.}

\item{d}{Smoothing parameter of the procedure for the backward}

\item{trace}{If TRUE then mars function prints details of the fitting process}
}
\description{
This function constructs a \code{mars.control} object that specifies
parameters used in the model fitting procedure.
}
