% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{Constructor for 'mars.control' objects}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{Maximum number of basis functions. The default value is is 2 and should be an even integer.}

\item{d}{The coefficient in the penalty term of the generalized cross validation measure. Default is 3.}

\item{trace}{Default is 'FALSE'.}
}
\value{
a 'mars.control. object
}
\description{
Constructor for 'mars.control' objects
}
\examples{
mc <- mars.control(Mmax = 10)
}
