% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{mars.control}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{Maximum number of basis functions. Should be an even integer. Default value is 2.}

\item{d}{Parameter d in GCV. Default value is 3.}

\item{trace}{Boolean to print details of the fitting process. Default value is FALSE.}
}
\value{
an object of class 'mars.control'
}
\description{
mars.control
}
