% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{helper for mars.control object}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{mmax}

\item{d}{d}

\item{trace}{trace}
}
\value{
returns an object of class mars.control
}
\description{
helper for mars.control object
}
\examples{
mc<-mars.control(Mmax=10,d=3,trace=FALSE)
mc
}
