% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{mars.control}
\usage{
S3 method for class 'mars.control'.
}
\arguments{
\item{Mmax}{Maximum number of basis functions. Should be an even integer.}

\item{d}{The coefficient in the penalty term of the generalized cross validation measure}

\item{trace}{Should we print status information about the fitting? Default is 'False'}
}
\value{
a list which class is "mars.control", and contain three variables: Mmax, d and trace
}
\description{
Constructor for 'mars.control' object. This function constructs a 'mars.control' object that specifies parameters used in the model fitting procedure.
}
\examples{
mc <- mars.control(Mmax = 10)
}
\seealso{
\link{plot.mars} for plotting the results

\link{predict.mars} for predictions.

\link{summary.mars}/\link{print.mars} for summarizing.
}
