% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{Constructor for 'mars.control' objects}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{Maximum number of basis functions. Should be an even integer.
Default value is 2.}

\item{d}{The parameter used in calculation of Generalized cross-validation.
Default value is 3.}

\item{trace}{A True or False value that allows user to see the candidate
basis functions that might reduce GCV in backward selection procedure similar
to linear regression modelling.}
}
\value{
an object of class 'mars.control', that is a list of parameters
}
\description{
This function constructs a 'mars.control' object that specifies
parameters used in mars.
}
\examples{
mc<-mars.control(Mmax=6,d=3,trace=FALSE)

}
