% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control = mars.control())
}
\arguments{
\item{formula}{object of formula to regress a response such as \code{y} and a predictor such as \code{x}. Use \code{.} for all in data}

\item{data}{input data frame}

\item{control}{mars.control object - choice of Mmax}
}
\value{
an object of class \code{mars}, inheriting from class \code{lm}. They return a list of class \code{lm} plus list of class `mars.

Some \code{lm} values:\if{html}{\out{<br>}}
\code{coefficients}.. holds coefficients on model\if{html}{\out{<br>}}
\code{rank}.......... rank of basis function\if{html}{\out{<br>}}
\code{residuals}..... error terms\if{html}{\out{<br>}}
\code{etc}........... additional \code{lm} values

additional to this list in \code{mars} is:\if{html}{\out{<br>}}
\code{formula}....... object of formula\if{html}{\out{<br>}}
\code{call}.......... call of formula matched\if{html}{\out{<br>}}
\code{y}............. data frame of response\if{html}{\out{<br>}}
\code{B}............. basis functions\if{html}{\out{<br>}}
\code{Bfuncs}........ optimize basis functions\if{html}{\out{<br>}}
\code{x_names}....... names of predictors\if{html}{\out{<br>}}
}
\description{
\code{mars} fits a model on data to capture a non-linear relationship. Using step-functions between knots, it splits basis functions and removes those that don't improve the model. Returning the optimal number of basis functions.
}
\details{
A fitted model using \code{mars}, primarily go through three stages. First when a call is made, it goes through a forward pass stepwise selection of basis functions in \code{fwd_stepwise()}, which creates splits on basis functions and optimizes these to find the best splits. Second, the results of fwd_stepwise, are passed to a backward stepwise selection of basis functions in \code{bwd_stepwise()}, which prunes basis functions by removing those that do not improve the model by checking Generalized Cross-Validation (GCV) and when GCV can no longer be reduced, we have our output. The final stage returns the results of the best basis functions, with their variable/predictor, sign and knot of their split. The output returned is augmented by additional parameters from \verb{mars.control().}
}
\examples{
set.seed(2026)

# Example 1:

m <- mars(y~., marstestdata,control = list(Mmax = 2L, d = 3, trace = FALSE))
summary(m)
print(m)
predict(m, newdata = marstestdata)
anova(m)
plot(m)

#Example 2: Spotify data, predicting popularity of songs
#data from: https://www.kaggle.com/datasets/maharshipandya/-spotify-tracks-dataset

fit <- mars(popularity ~. , spotify, control = mars.control(Mmax=4))
summary(fit)
anova(fit)

# Example 3: Airbnb prediction on rental price
## dataset from: https://www.kaggle.com/datasets/dgomonov/new-york-city-airbnb-open-data/

airbnb <- na.omit(airbnb)

fit <- mars(price ~. , airbnb, control = mars.control(Mmax=4))
summary(fit)
anova(fit)
plot(fit)

# Example 4: Pima Indians Diabetes (Pima) dataset, predict chance of diabetes for women
# data from: https://machinelearningmastery.com/standard-machine-learning-datasets-for-imbalanced-classification/

fit <- mars(X0.627 ~. , diabetes, control = mars.control(Mmax=4))
summary(fit)
predict(fit, newdata = diabetes)
anova(fit)
plot(fit)

}
\references{
Jerome H. Friedman "Multivariate Adaptive Regression Splines," The Annals of Statistics, Ann. Statist. 19(1), 1-67, (March, 1991)
}
\seealso{
\code{\link[=print.mars]{print}},
\code{\link[=summary.mars]{summary}},
\code{\link[=predict.mars]{predict}},
\code{\link[=plot.mars]{plot}}
}
\author{
Zvikomborero Kennedy Jokonya | email: zkj@sfu.ca |
}
