% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control = mars.control())
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame containing the a response variable and predictors}

\item{control}{a mars.control object created using mars.control()}
}
\value{
An S3 model of class "mars"
}
\description{
Fit Friedman's Multivariate Adaptive Regression Splines (MARS) model.
}
\examples{
## Analyzing Boston value data found in the Boston dataset

mars.mod <- mars(medv~zn+crim+rm,data=Boston,control=mars.control())
print(mars.mod)
}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, 19(1),
 1–67. https://doi.org/10.1214/aos/1176347963
}
\seealso{
\describe{
\item{anova.mars}{}
\item{plot.mars}{}
\item{predict.mars}{}
\item{print.mars}{}
\item{summary.mars}{}
}
}
