% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control=NULL, ...)
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame containing the data}

\item{control}{an object of class 'mars.control'}

\item{...}{further arguments}
}
\value{
an S3 object of class mars that includes the final regression model and a description of the basis functions which are constructed by the hinge functions
}
\description{
MARS (Multivariate Adaptive Regression Splines) is a regression analysis that automatically splits and better linear fit into non-linear models.
}
\details{
MARS - forward stepwise algorithm:
Forward stepwise fits the data nicely by adjusting the coefficient values as well as deriving a proper set of basis functions. Forward stepwise produces basis functions which do not have zero pairwise product expectations. The advantage of this is that basis functions, except for B0, can be removed without leaving a hole in the predictor space.

MARS - backward stepwise algorithm:
Jstar(J*) backward stepwise is all of the basis function set and derived from forward stepwise. One basis function is deleted at each iteration of the outer loop, while the inner loop chooses which one to be deleted. The chosen basis function to be deleted is either the one improving the model by its removal or degrading the model the least by the removal. Constant basis function B0(x) = 1 is not considered for elimination. A sequence of (Mmax - 1) models is what is constructed from backward stepwise, and the current sequence has one less basis function than that of the previous sequence. Once the iteration is terminated, the users are left with the best model.
}
\examples{
test <- mars(y~x1+x2, data=mars::marstestdata, control=mars.control(Mmax=2))
}
\references{
Jerame H. Friedman. "Multivariate Adaptive Regression Splines."
Ann, Statist. 19 (1) 1 - 67, March, 1991 \url{https://doi.org/10.1214/aos/1176347963}
}
\seealso{
\link{plot.mars} for plotting the basis functions of the mars object

\link{predict.mars} for making predictions on a new data using the mars object

\link{summary.mars} for summarizing the mars object

\link{print.mars} for printing the mars object
}
\author{
Siyul Sam Byun, So Yeon Park, Dawu Liu
}
\concept{methods for mars object}
