% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Title}
\usage{
mars(formula, data, control = NULL, ...)
}
\arguments{
\item{formula}{a R formula}

\item{data}{a data}

\item{control}{a mars.control object}
}
\value{

}
\description{
Title
}
