% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control)
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame containing the data for the model}

\item{control}{an object of class 'mars.control'}
}
\value{
an object of class 'mars' which will be used for plot, predict, etc.
}
\description{
Fit Friedman's Multivariate Adaptive Regression Splines (MARS) model.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))

}
