% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control) #for generating a mars object
}
\arguments{
\item{formula}{an R formula specifying the dependent and independent variables in the model. The formula should take the form y ~ x1 + x2 + ..., where y is the response variable and x1, x2, etc. are the predictor variables.}

\item{data}{a data frame containing the variables in the formula.}

\item{control}{an object of class mars.control that specifies parameters used in the model fitting procedure. By default, the control object is constructed using \code{mars.control()}, with default values for the parameters.}
}
\value{
an object of class 'mars' which will be used for plot, predict, etc.
}
\description{
Fit Friedman's Multivariate Adaptive Regression Splines (MARS) model.
}
\details{
The function first extracts the response variable and the predictor variables
from the data frame using the formula and \verb{model.frame() function}. Then, it performs
forward stepwise regression on the data using the \code{fwd_stepwise()} function, followed by
backward stepwise regression on the selected terms using the \code{bwd_stepwise()} function.
The resulting basis functions and basis matrix are stored in the bwd object. Finally, a
linear model is fit to the data using the \code{lm()} function, with the basis matrix and response
variable as inputs. The resulting object is returned as a list with class 'mars'.
}
\examples{
#example 1
# mars() with ISLR::Wage dataset
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
# Show results
fit.mars

#example 2
# create your own dataset
n <- 1000
x1 <- rnorm(n)
x2 <- rnorm(n)
x3 <- rnorm(n)
y <- 2*x1 + 3*x2 + 4*x3 + rnorm(n)
df <- data.frame(x1=x1,x2=x2,x3=x3,y=y)

# fit a MARS model
fit <- mars(y ~ x1 + x2 + x3, data=df)
#show results
fit

#example 3
# another example using data(iris)
fit.iris <- mars(Sepal.Length ~., data=iris, control = mars.control(Mmax=10))
# Show results
fit.iris


}
\references{
Friedman, J. H. (1991). Multivariate adaptive regression splines. The Annals of Statistics, 19(1), 1-67.
}
\seealso{
\code{\link[=print.mars]{print.mars()}}

\code{\link[=summary.mars]{summary.mars()}}

\code{\link[=predict.mars]{predict.mars()}}

\code{\link[=plot.mars]{plot.mars()}}

\code{\link[=anova.mars]{anova.mars()}}
}
\author{
Group DJT: Dustin Poon, James Lee, Tyler Oh
}
