% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control = NULL)
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame containing your data}

\item{control}{an optional object of class 'mars.control'}
}
\value{
an object of class 'mars' which can be passed to plot,predict,summary and print methods.
}
\description{
Fit Friedman's Multivariate Adaptive Regression Splines (MARS) model.
}
\details{
The MARS function attempts to fit non-linear data as accurately as possible. To do this the function defines one or more hinge-points that act as connectors between two or more linear functions.This process is completed through a multistage process including a forward pass and backwards pass.

The forward pass algorithm takes a formula and data as input and attempts to output a matrix of basis functions of size Mmax. To do this, the algorithm considers each data point for each predictor and generates candidate basis function pairs. These candidates are selected based on reduction to the model's residual sum-of-squares.
The basis function matrix contains a row for each observations and a column for each basis function.

The forward pass identifies every basis function that reduces the model's residual sum-of-squares, but this can produce an over-fit model. To create a model that generalized better with new data, the matrix is handed to the backwards pass algorithm. The algorithm 'prunes' any basis function that adds no significant predictive accuracy to the overall model.
The backwards algorithm compares subsets of basis function using generalized cross-validation scores in which 'd' controls the amount of penalization there is for extra model terms. By tracking GCV scores for each possible combination of basis function, the best possible model can be selected.
The backwards pass algorithm updates the basis matrix to contain only the function from the optimal model.
}
\examples{
mm<-mars(y~.,dat=mars::marstestdata)
}
\references{
Jerome H. Friedman. Multivariate Adaptive Regression Splines (with discussion).Annals of Statistics 19/1, 1991. \url{https://statistics.stanford.edu/research/multivariate-adaptive-regression-splines.}
}
\seealso{
\link{mars.control} for constructing control objects

\link{plot.mars} for plotting results

\link{predict.mars} for predictions

\link{summary.mars} for summarizing mars objects

\link{print.mars} for printing mars objects
}
\author{
Tauseef Kashtwari, Promit Chowdhury, Ibraheem Azad
}
