% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Fit a MARS model}
\usage{
mars(formula, data, control)
}
\arguments{
\item{formula}{An R formula}

\item{data}{A data frame containing the data for the model}

\item{control}{An object of class \code{mars.control}}
}
\value{
An object of class \code{mars} that includes the final regression and a description
of the basis functions. There are plot, predict, summary and print methods for mars object.
}
\description{
Build a regression model using the techniques in Friedman's papers "Multivariate Adaptive Regression Splines (MARS)".
}
\details{
The algorithm will search for, and discover, nonlinearities in the data that help maximize predictive accuracy.
Multivariate adaptive regression splines (MARS), an algorithm that essentially creates a piecewise linear model
which provides an intuitive stepping block into nonlinearity after grasping the concept of linear regression
and other intrinsically linear models.
}
\examples{
mm <- mars(y~., data=mars::marstestdata)
}
\references{
Jerome H. Friedman. “Multivariate Adaptive Regression Splines.”
Ann. Statist. 19(1) 1 - 67, March, 1991. https://doi.org/10.1214/aos/1176347963.
}
\seealso{
\code{\link[=mars.control]{mars.control()}} for constructing the control object \cr
\code{\link[=plot.mars]{plot.mars()}} for plotting the results \cr
\code{\link[=predict.mars]{predict.mars()}} for predictions \cr
\code{\link[=summary.mars]{summary.mars()}} for summarizing mars objects \cr
\code{\link[=print.mars]{print.mars()}} for printing mars objects

Other method: 
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\author{
Tauseef Kashtwari, Promit Chowdhury, Ibraheem Azad
}
\concept{method}
