% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control = NULL)
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame}

\item{control}{an object of class 'mars.control'}
}
\value{
an object of mars which has the model
}
\description{
Fit Friedman's Multivariate Adaptive Regression Splines (MARS) model.
}
\details{
The model we receive at the end from the mars algorithm is a many basis functions in a linear combination, found by least squares
}
\examples{
mm<-mars(y~.,dat=mars::marstestdata)
}
\author{
Ajay Behal Lei Liu Linden Gueck
}
