% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control = mars.control())
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame containing the data}

\item{control}{an object of class 'mars.control'}
}
\value{
an object of class 'mars.control'
}
\description{
Fit Friedman's Multivariate Adaptive Regression Splines (MARS) model.
}
\details{
mars
}
