% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARS.R
\name{MARS}
\alias{MARS}
\title{Estimation of a sequence of precision matrices (MARS)}
\usage{
MARS(
  X = NULL,
  stoptol = 1e-04,
  Lambdapath = seq(1, 0.5, -0.1),
  maxiter = 10,
  stopmethod = "fix",
  printmain = FALSE,
  printsub = FALSE,
  sigma = 1,
  fixnumber = 10,
  maxlambdacheck = TRUE
)
}
\arguments{
\item{X}{A p times n sample matrix.}

\item{stoptol}{A small value to control the stop tolerance.}

\item{Lambdapath}{A vector of tuning parameters.}

\item{maxiter}{A integer is set to be the maximum iteration number for the
main loop.}

\item{stopmethod}{A string to control the stop method, could be "bigs" or
"fix". If it is "bigs", then the algorithm will stop when the number of
non-zero off-diagonal components exceeds n times fixnumber. If it is "fix",
then the algorithm will only calculate the estimator with first fixnumber
biggest tuning parameters.}

\item{printmain}{A boolean for controlling whether to print the main results
during the calculation or not.}

\item{printsub}{A boolean for controlling whether to print the inner problem
results during the calculation or not.}

\item{sigma}{The initial value of parameter sigma.}

\item{fixnumber}{As described in the previous.}

\item{maxlambdacheck}{A boolean to decide whether to narrow the lambda from
its upper bound path or not.}
}
\value{
A list of solutions: Omegapath, Lambdapath, and timepath, which are
  the estimated solutions, the tunning parameters, and the computation time
  respectively.
}
\description{
Under the high-dimensional setting, this function is designed
  for generating a solution path of the precision matrices by a second-order
  algorithm with the l1-penalized D-trace loss.
}
\details{
Input a p times n sample matrix and a vector of tuning parameters to
  obtain a path of estimated precision matrices, where p is the sample
  dimension and n is the sample size. Note that our algorithm is designed for
  the case that p is much greater than n.
}
\references{
Qian Li, Binyan Jiang, Defeng Sun. MARS: A second-order reduction
  algorithm for high-dimensional sparse precision matrices estimation.
}
