% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control = NULL, ...)
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame containing your data}

\item{control}{an object of class 'mars.control'}

\item{...}{unused}
}
\value{
an object of class 'mars'
}
\description{
Multivariate Adaptive Regression Splines (MARS)
}
\examples{
 mm <- mars(wage ~ age,dat=ISLR::Wage)
}
