% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines}
\usage{
mars(formula, data, control = NULL, ...)
}
\arguments{
\item{formula}{an R formula to perform the regression on}

\item{data}{data set containing the data to perform mars on}

\item{control}{object of class mars.control}

\item{...}{additional argument to pass to mars()}
}
\value{
the mars() would return an object of class 'mars' that includes the basis function obtained through final regression. the mars object has plot,predict and summary methods for mars object.
}
\description{
Mars package run a step wise selection algorithm to look for clusters in a data and return a basis function. There are multiple methods like print, summary and plot to visualizing the mars object
}
\details{
call the mars function as follows:-
}
\examples{
 mars(wage~age,data=ISLR::Wage)
}
\references{
Jerome H Friedman. "Multivariate Adaptive Regression Splines"
Ann. Statist. 19 (1) 1-67,March, 1991. \url{https://doi.org/10.1214/aos/1176347963}

Brad McNeney STAT360 Advanced R for Data Science
}
\seealso{
\link{plot.mars} for plotting the results

\link{summary.mars} and \link{print.mars} for summarizing and printing mars object

\link{predict.mars} to obtain prediction values
}
\author{
Simarprit Singh
}
