% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{MARS}
\usage{
mars(formula, data, control)
}
\arguments{
\item{formula}{interface to specify response and explanatory variables, a variable of the model to be fitted.}

\item{data}{A data frame, list, or environment, containing the variables in the model. If not found in the data, the variables are taken from the environment of formula}

\item{control}{An object of mars.control with Mmax = 2 and d = 3, by default.}
}
\value{
a \code{mars} object, which contain the model
}
\description{
The mars() function, which is the main function in the MARS Package, performs a Multivariate Adaptive Regression Splines (MARS) analysis. The MARS algorithm builds a regression model by partitioning the predictor variables into segments and fitting a linear regression model to each segment.
}
\details{
MARS is a non-parametric regression technique that partitions the predictor space into simple regions defined by piecewise linear functions of the input variables. The MARS algorithm uses a forward stepwise approach to iteratively add basis functions to the model.
}
\examples{
test1 <- mars(y~., data1[1:100,], mars.control(Mmax=6))
print(test1)
summary(test1)
plot(test1)
anova(test1)
predict(test1,data1[101:200,1:10])

}
\references{
Jerome H. Friedman. Multivariate Adaptive Regression Splines (with discussion).Annals of Statistics 19/1, 1991. \url{https://statistics.stanford.edu/research/multivariate-adaptive-regression-splines.}
}
\seealso{
\link{mars.control} for constructing control objects, The default mars.control has Mmax is 2 and d is 3.

\link{plot.mars} for generates four different residual plots.

\link{predict.mars} for generates new dependent values based on the dataset.

\link{summary.mars} for provides a summary table of the estimated coefficients of all the independent variables with respect to the dependent variable.

\link{print.mars} for printing the coefficients for each independent variable.

\link{anova.mars} for compute analysis of variance (or deviance) tables for fitted \code{mars} model objects.
}
