% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{Multivariate Adaptive Regression Splines (MARS)}
\usage{
mars(formula, data, control = mars.control())
}
\arguments{
\item{formula}{an R formula}

\item{data}{a data frame containing the data for the model}

\item{control}{an object of class 'mars.control'}
}
\value{
an object of class 'mars' that include the final regression
and a description of the basis functions. There are plot, predict, anova,
summary and print methods for mars objects.
}
\description{
Mars is a form of stepwise linear regression to improve
performance of the model in the regression setting.
}
\details{
The MARS function is applied for constructing Multivariate
Adaptive Regression Splines models. It refers to a stepwise linear
regression format that produces the required model. As it proceeds
it is ensured that for example, it will first look for the single
point across the range of X values where two different linear relationships
between Y and X achieve the smallest error via GCV computation. The "mars"
function takes as input a formula representing the relationship between the
response and explanatory variables, the data set to fit the model on and a
control object. It returns an object of class "mars". This returned object
is later used for further output results on different methods.

The mars function comprises several methods, including anova.mars,
plot.mars, predict.mars, print.mars, and summary.mars. These methods
generate outputs through the application of the corresponding
techniques - anova, plot, predict, print, and summary.
}
\examples{
library(ISLR)
data("Auto")
data("College")
data("Hitters")
# example1:
Auto_mars<- mars(Auto$mpg~Auto$displacement+Auto$weight, data=Auto)
# example2:
College_mars=mars(College$Grad.Rate~College$F.Undergrad+College$P.Undergrad,data=College)
# example3:
Hitters_mars=mars(Hitters$Salary~Hitters$HmRun+Hitters$Hits, data=Hitters)

}
\references{
Friedman, J. H. (1991). Multivariate Adaptive Regression Splines.
The Annals of Statistics, 19(1), 1–67. http://www.jstor.org/stable/2241837

Boehmke, B. (2020, February). Chapter 7 Multivariate Adaptive Regression
Splines | Hands-On Machine Learning with R.
Github.io. https://bradleyboehmke.github.io/HOML/mars.html
}
\seealso{
anova, predict, print, plot, summary
}
\author{
Jiawei Lin, Yanning Chen, Isabah Tabenda Hasan
}
