% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{mars: Multivariate Adaptive Regression Splines Model}
\usage{
S3 method for class 'mars'
}
\arguments{
\item{formula}{data frame for formula}

\item{data}{a data frame containing the data}

\item{control}{an object of class 'mars.control'}
}
\value{
an object of class 'mars' that contains the value: fitted value and coefficients.
}
\description{
Build a regression model using the techniques in Friedman's papers "Multivariate Adaptive Regression Splines"
Mars can be used to model onlinear relationships between a set of predictor variables and a response variable.
}
\details{
an object of class 'mars.control'that includes the final regression and a description of the basis function.There are plot,predit, summary and print methods for mars objects.
}
\examples{
mm <- mars(y~., dat=mars::marstestdata)
}
\references{
Jerome H. Friedman The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67. \url{https://doi.org/10.1214/aos/1176347963}.
}
\seealso{
\link{mars.control} for constructing the control object.

\link{plot.mars} for plotting the results, \link{predict.mars} for predictions, \link{summary.mars}/\link{print.mars} for summarizing

Other methods: 
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{summary.mars}()}
}
\author{
Yanrong Shi, Yongliang Ye
}
\concept{methods}
