% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars}
\alias{mars}
\title{MARS}
\usage{
mars(formula, data, control)
}
\arguments{
\item{formula}{an object of class \code{formula} (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{control}{an object of class \code{mars.control}. The default mars.control has Mmax (the maximum number of model terms) is 2 and d (the smoothing parameter) is 3, but users can customize their mars.control.}
}
\value{
a \code{mars} object, which contain the model
}
\description{
The mars() function, which is the main function in the MARS Package, performs a Multivariate Adaptive Regression Splines (MARS) analysis. The MARS algorithm builds a regression model by partitioning the predictor variables into segments and fitting a linear regression model to each segment.
}
\details{
MARS is a non-parametric regression technique that partitions the predictor space into simple regions defined by piecewise linear functions of the input variables. The MARS algorithm uses a forward stepwise approach to iteratively add basis functions to the model. This is done until a specified maximum number of functions or a stopping criterion is reached. Then, a backward stepwise approach is used to prune the modelby removing any basis functions that do not contribute significantly to the model’s performance.
}
\examples{
test <- mars(y~.,data=data1[1:100,])

print(test)
summary(test)
plot(test)
anova(test)
predict(test,data1[101:200,])

}
\references{
Jerome H. Friedman. Multivariate Adaptive Regression Splines (with discussion).Annals of Statistics 19/1, 1991. \url{https://statistics.stanford.edu/research/multivariate-adaptive-regression-splines.}
}
\seealso{
\link{mars.control} for constructing control objects, The default mars.control has Mmax (the maximum number of model terms) is 2 and d (the smoothing parameter) is 3, but users can customize their mars.control.

\link{plot.mars} for generates four different residual plots。

\link{predict.mars} for generates new dependent values based on the dataset.

\link{summary.mars} for provides a summary table of the estimated coefficients of all the independent variables with respect to the dependent variable.

\link{print.mars} for printing the coefficients for each independent variable.
}
