% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-s4-class.R
\docType{class}
\name{MARSdata-class}
\alias{MARSdata-class}
\title{MARSdata S4 object}
\description{
S4 class that organizes the various data inputs for the MARS model. \code{MARSdata} simply inherits the slots from 6 component classes:
\code{Dmodel}, \code{Dstock}, \code{Dfishery}, \code{Dsurvey} \code{DCKMR}, and \code{Dtag}, where the \code{D}- prefix denotes an object for model data.
}
\details{
For convenience, most arrays and matrices have the associated dimensions in the variable name. For example, \code{Cobs_ymfr} represents
observed catch with the dimension following the underscore, following this template:

\tabular{ll}{
\code{y} \tab Year \cr
\code{m} \tab Season \cr
\code{a} \tab Age \cr
\code{r} \tab Region \cr
\code{f} \tab Fishery \cr
\code{i} \tab Index \cr
\code{s} \tab Stock
}
}
\section{Slots}{

\describe{
\item{\code{Dmodel}}{Class \linkS4class{Dmodel} containing parameters for model structure (number of years, ages, etc.)}

\item{\code{Dstock}}{Class \linkS4class{Dstock} containing stock parameters (growth, natural mortality, etc.)}

\item{\code{Dfishery}}{Class \linkS4class{Dfishery} containing fishery data (catch, size and stock composition, etc.)}

\item{\code{Dsurvey}}{Class \linkS4class{Dsurvey} containing survey data (indices of abundance)}

\item{\code{DCKMR}}{Class \linkS4class{DCKMR} containing genetic close-kin data}

\item{\code{Dtag}}{Class \linkS4class{Dtag} containing tagging data}

\item{\code{Misc}}{List for miscellaneous inputs as needed}
}}

\section{Slots inherited from Dmodel}{

\describe{
\item{\code{ny}}{Integer, number of years}
\item{\code{nm}}{Integer, number of seasons}
\item{\code{na}}{Integer, number of ages}
\item{\code{nl}}{Integer, number of length bins. Set to zero if lengths are not modeled.}
\item{\code{nr}}{Integer, number of spatial regions}
\item{\code{ns}}{Integer, number of stocks}
\item{\code{lbin}}{Vector of lower boundary of length bins. Length \code{nl + 1}}
\item{\code{lmid}}{Vector of midpoint of length bins. Length \code{nl}}
\item{\code{Fmax}}{Numeric, maximum allowable instantaneous fishing mortality rate (units of per season). Defaults to 3.}
\item{\code{nitF}}{Integer, number of iterations to solve Baranov catch equation from observed catch. Defaults to 5.}
\item{\code{dist_type}}{Character. Whether to estimate seasonal stock distribution with a movement matrix \code{"mov"} or re-distributing total abundance by a distribution vector \code{"dist"}}
\item{\code{y_phi}}{Integer, the year from which to obtain values of natural mortality and fecundity for the unfished stock-recruit replacement line (\code{phi}). Relevant if natural mortality or fecundity are time-varying. Defaults to 1.}
}
}

\section{Slots inherited from Dstock}{

\describe{
\item{\code{len_ymas}}{Length-at-age. Only needed if \code{Dmodel@nl > 0}.}
\item{\code{sdlen_ymas}}{Standard deviation in length-at-age}
\item{\code{LAK_ymals}}{Length-at-age probability array. If empty, values will be calculated in \link{check_data} (see \link{calc_LAK}).}
\item{\code{mat_yas}}{Proportion mature by age class}
\item{\code{fec_yas}}{Fecundity, i.e., spawning output, of mature animals}
\item{\code{swt_ymas}}{Stock weight-at-age}
\item{\code{Md_yas}}{Natural morality. Ignored if M is estimated.}
\item{\code{m_spawn}}{Integer, season of spawning}
\item{\code{m_rec}}{Integer, season of recruitment}
\item{\code{SRR_s}}{Character vector of stock-recruit relationship by stock. See \link{calc_recruitment} for options.}
\item{\code{delta_s}}{Fraction of season that elapses when spawning occurs}
\item{\code{natal_rs}}{Boolean matrix that indicates whether stock \code{s} spawns when in region \code{r} at time of spawning.}
}
}

\section{Slots inherited from Dfishery}{

\describe{
\item{\code{nf}}{Integer, number of fleets}
\item{\code{Cobs_ymfr}}{Total fishery catch}
\item{\code{fwt_yamfs}}{Fishery weight at age. Set to 1 when fleet catch is in units of abundance. Set to stock weight at age by default.}
\item{\code{CAAobs_ymafr}}{Fishery catch at age composition}
\item{\code{CALobs_ymlfr}}{Fishery catch at length composition}
\item{\code{fcomp_like}}{Character, likelihood for the fishery composition data. See \code{type} argument in \link{like_comp} for options}
\item{\code{CAAN_ymfr}}{Sample size of the catch at age vector by season if using the multinomial or Dirichlet-multinomial likelihoods}
\item{\code{CALN_ymfr}}{Sample size of the catch at length vector by season if using the multinomial or Dirichlet-multinomial likelihoods}
\item{\code{CAAtheta_f}}{Catch at age dispersion parameter if using the Dirichlet-multinomial likelihood. Default set to 1.}
\item{\code{CALtheta_f}}{Catch at length dispersion parameter if using the Dirichlet-multinomial likelihood. Default set to 1.}
\item{\code{sel_block_yf}}{Index of dummy fleets to model time blocks of selectivity}
\item{\code{sel_f}}{Character matrix of functional form for selectivity}
\item{\code{SC_ymafrs}}{Stock composition data.}
\item{\code{SC_aa}}{Boolean matrix that groups together age classes for the stock composition data}
\item{\code{SC_ff}}{Boolean matrix that groups together fleets for the stock composition data}
\item{\code{SC_like}}{Character, likelihood for the stock composition data. See \code{type} argument of \link{like_comp} for options}
\item{\code{SCN_ymafr}}{Sample size of the stock composition vector if using the multinomial or Dirichlet-multinomial likelihoods}
\item{\code{SCtheta_f}}{Stock composition dispersion parameter if using the Dirichlet-multinomial likelihood. Default set to 1.}
\item{\code{SCstdev_f}}{Stock composition standard deviation if using the lognormal likelihood. Default set to 0.1.}
}
}

\section{Slots inherited from Dsurvey}{

\describe{
\item{\code{ni}}{Integer, number of indices of abundance. Zero is possible.}
\item{\code{Iobs_ymi}}{Observed indices}
\item{\code{Isd_ymi}}{Lognormal standard deviation of the observed indices}
\item{\code{unit_i}}{Character vector, units of the index. Set to \code{"B"} to use stock weight at age (default) or \code{"N"} for abundance (numbers).}
\item{\code{IAAobs_ymai}}{Survey age composition}
\item{\code{IALobs_ymli}}{Survey length composition}
\item{\code{icomp_like}}{Character, likelihood for the composition data. See \link{like_comp} for options}
\item{\code{IAAN_ymi}}{Sample size of the index age composition by season if using the multinomial or Dirichlet-multinomial likelihoods}
\item{\code{IALN_ymi}}{Sample size of the index length composition by season if using the multinomial or Dirichlet-multinomial likelihoods}
\item{\code{IAAtheta_i}}{Index age composition dispersion parameter if using the Dirichlet-multinomial likelihood}
\item{\code{IALtheta_i}}{Index length composition dispersion parameter if using the Dirichlet-multinomial likelihood}
\item{\code{samp_irs}}{Boolean array that specifies the regions and stocks sampled by the index. \code{samp[i, r, s]} indicates whether index \code{i} operates in region \code{r} and catches stock \code{s}.}
\item{\code{sel_i}}{Character matrix of functional forms for selectivity}
\item{\code{delta_i}}{The elapsed fraction of time in the seasonal time step when the index samples the population.}
}
}

\section{Slots inherited from DCKMR}{

\describe{
\item{\code{POP_s}}{A list by stock of data frames for parent-offspring pairs. Each row in the data frame corresponds to a "sampling unit" defined by the columns:
\tabular{ll}{
\code{a} \tab Capture year of parent \cr
\code{t} \tab Age at capture of parent \cr
\code{y} \tab Birth year of offspring \cr
\code{n} \tab Number of pairwise comparisons \cr
\code{m} \tab Number of POPs \cr
}
}
\item{\code{HSP_s}}{A list by stock of data frames for half-sibling pairs. Each row in the data frame corresponds to a "sampling unit" defined by the columns:
\tabular{ll}{
\code{yi} \tab Birth year of older sibling \cr
\code{yj} \tab Birth year of younger sibling \cr
\code{n} \tab Number of pairwise comparisons \cr
\code{m} \tab Number of HSPs
}
}
\item{\code{CKMR_like}}{Character, likelihood for the POP and HSP sampling units. See \link{like_CKMR} for options.}
}
}

\seealso{
\link{MARSdata-class} \code{\link[=check_data]{check_data()}} \link{Dmodel-class} \link{Dstock-class} \link{Dfishery-class} \link{Dsurvey-class} \link{DCKMR-class} \link{Dtag-class}
}
\keyword{MARSdata}
