% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marstestdata}
\alias{marstestdata}
\title{Test data set for Multivariate Adaptive Regression Splines (MARS) function}
\format{
A data frame with 100 rows and 11 columns.
}
\usage{
data(marstestdata)
}
\description{
A data set containing one response variable "y" and ten explanatory variables
"x1" to "x10" made to test the "mars" function. This data set is written to run the examples
shown in the documentation of MARS and method predict.mars.
}
\details{
\itemize{
\item y. response variable.
\item x1:x10. ten explanatory variables
}
}
\author{
Dr. Brad McNeney \email{brad_mcneney@sfu.ca}
}
\keyword{datasets}
