% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marstestdata}
\alias{marstestdata}
\title{A test dataset for the mars package.}
\format{
A data frame with 100 rows and 11 variables:
\describe{
\item{y}{response variable}
\item{x1}{explanatory variable}
\item{x2}{explanatory variable}
\item{x3}{explanatory variable}
\item{x4}{explanatory variable}
\item{x5}{explanatory variable}
\item{x6}{explanatory variable}
\item{x7}{explanatory variable}
\item{x8}{explanatory variable}
\item{x9}{explanatory variable}
\item{x10}{explanatory variable}
}
}
\usage{
marstestdata
}
\description{
A dataset of size N=100 with n=10 explanatory variables,
and a response variable that depends on only the first two
explanatory variables.
}
\keyword{datasets}
