% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{new_mars.control}
\alias{new_mars.control}
\title{Constructor for \code{mars.control} objects}
\usage{
new_mars.control(Mmax, d, trace)
}
\arguments{
\item{Mmax}{Maximum number of basis functions. Should be an even interger Default value is 1.}

\item{d}{The coefficient in the penalty term of the generalized cross validation measure. Default is 3.}

\item{trace}{Should we print information about the fitting? Default is \code{FALSE}}
}
\value{
a \code{mars.control} object
}
\description{
This function constructs a \code{mars.control} object that specifies
parameters used in the model fitting procedure
}
\examples{
mc<-mars.control(Mmax=10)
}
