% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot Diagnostics for a mars Object}
\usage{
\method{plot}{mars}(mars, which = c(1, 2, 3, 4), col = "purple", numpoints = 3, ...)
}
\arguments{
\item{mars}{a mars object, typically the result of [mars()]}

\item{which}{a subset of the numbers 1:4 specifying which plots to create, by default 1:4
\describe{
 \item{1. "CDF" plot of absolute value of residuals}{}
 \item{2. "Residual vs Fitted" plot}{}
 \item{3. "Normal Q-Q" plot}{}
 \item{4. "Scale-Location" plot}{}
 }}

\item{col}{color for: the line in the cdf plot or the points n the remaining points}

\item{numpoints}{how many of the largest absolute value points to plot}

\item{...}{Extra parameters to pass to plotting functions}
}
\description{
Four plots (selected by which) are available: a cdf plot of |residuals|, a fitted vs residual plot,
a q-q plot and a scale-location plot
}
\examples{
## Analysis of the car data found in mtcars
## Printing the cdf, fitted vs residual and q-q plot
## with 4 label points and green points

m<-mars(form=mpg~.,data=mtcars,control=mars.control())
plot(m,which=c(1,2,3),col="green",numpoints=4)

}
\references{
https://CRAN.R-project.org/package=earth
}
\seealso{
Other methods: 
\code{\link{anova.mars}()},
\code{\link{predict.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
