% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot a mars object}
\usage{
\method{plot}{mars}(x, ...)
}
\arguments{
\item{x}{a mars object}

\item{...}{further arguments}
}
\description{
Plot basis functions that are constructed with a single explanatory variable
or two explanatory variables
}
\examples{
mar <- mars(y~x1+x2+x3, data=mars::marstestdata, control=mars.control(Mmax=6))
plot(x=mar)
}
