% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot a MARS Model}
\usage{
\method{plot}{mars}(x, ...)
}
\arguments{
\item{x}{MARS model object (input mars object)}

\item{...}{further arguments passed to or from other methods}
}
\value{
2D plot for single variable basis function
3D plot for double variable basis function
}
\description{
Plot of the fitted basis functions.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
plot.mars(fit.mars)
}
\seealso{
\code{\link[=mars]{mars()}}

\code{\link[=print.mars]{print.mars()}}

\code{\link[=summary.mars]{summary.mars()}}

\code{\link[=predict.mars]{predict.mars()}}

\code{\link[=anova.mars]{anova.mars()}}
}
