% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mars}
\alias{plot.mars}
\title{Plotting Multivariate Adaptive Regression Splines (MARS) fit}
\usage{
\method{plot}{mars}(x, ...)
}
\arguments{
\item{x}{an object of class "mars".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
There is no return value for this function. This function produces a plot for mars x.
}
\description{
plot method for class "mars".
}
\details{
plot.mars method produces plots for the input mars object. It is limited to three dimensional plots, therefore
it only produces plots for basis functions with one or two variables in them. NO PLOT IS PRODUCED FOR BASIS FUNCTIONS
WITH MORE THAN TWO VARIABLES. This method identifies the basis functions with 1 or 2 components from Bfuncs list of mars object.
Index for one and two components basis function is stored separately in two different vectors. Then it splits the plot into a matrix of
plots based on total number of one and two components basis functions. "plot" function is then used to plot all the basis functions with
one variable in it, and "persp" function is used to plot basis functions with two variable components in it. Graphical parameters are reset
at the end using an exit handler.
}
\examples{
plot(testmars)
}
\references{
Friedman, J.H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, Vol. 19, No. 1 (Mar., 1991), pp. 1-67
}
\seealso{
"Multivariate Adaptive Regression Splines (MARS)" to create an x of class "mars", summary.mars method to summarise mars xs, predict.mars method for prediction, and print.mars to print mars xs.
}
\author{
Samir Arora and Gurpal Singh Tulli.
}
