% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot a mars object}
\usage{
\method{plot}{mars}(x, ...)
}
\arguments{
\item{x}{a mars x}

\item{...}{additional arguments to pass to plot()}
}
\description{
Plots the fitted basis functions that depend on
explanatory variable(main effects) or two explanatory variables
(two-way interactions). Use \code{predict.lm()} to see the residual plots
}
\examples{
mm<-mars(y~x1+x2,data=marstestdata,mars.control(Mmax=4))
plot(mm,col="red")
class(mm)<-"lm";plot(mm) #Standard diagnostic plots

}
\seealso{
Other methods: 
\code{\link{predict.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
