% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mars}
\alias{plot.mars}
\title{plot.mars}
\usage{
\method{plot}{mars}(OBJECT, ...)
}
\arguments{
\item{OBJECT}{a mars object}

\item{...}{these are additional arguments for plotting mars object}
}
\description{
plot.mars
}
\examples{
mm <- mars(y~x1+x2,data=marstestdata,mars.control(Mmax=4))
}
\author{
Tauseef Kashtwari, Promit Chowdhury, Ibraheem Azad
}
