% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot method for the mars object}
\usage{
\method{plot}{mars}(x, ...)
}
\arguments{
\item{x}{A \code{mars} object}

\item{...}{Additional arguments for plotting}
}
\value{
Four diagnose plots:
\itemize{
\item Residuals vs Fitted value
\item Normal Q-Q plot
\item Squared standardized residuals vs. Fitted value
\item Cook's distance.
}
}
\description{
Plots the fitted basis functions that depend on
explanatory variable(main effects) or two explanatory variables
(two-way interactions).
}
\examples{
mm <- mars(y~., data=mars::marstestdata)
plot(mm)
}
\seealso{
Other method: 
\code{\link{mars}()},
\code{\link{predict.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{method}
