% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{plot.mars}
\alias{plot.mars}
\title{plot various aspects of the mars object}
\usage{
\method{plot}{mars}(object)
}
\arguments{
\item{object}{an mars object}
}
\value{
plots of residuals vs fitted, qqplot, Scale-location, and Residual leverage
}
\description{
plot various aspects of the mars object
}
\examples{
# see forest fire ex. in testfile
plot.mars(object)

}
