% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot mars}
\usage{
\method{plot}{mars}(OBJECT, ...)
}
\arguments{
\item{OBJECT}{a mars object}

\item{description}{plots the fitted basis function}
}
\description{
Plot mars
}
\examples{
mm <- mars(y~x1+x2,data=marstestdata,mars.control(Mmax=4))
}
\author{
Ajay Behal Lei Liu Linden Gueck
}
