% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{plot.mars method}
\usage{
\method{plot}{mars}(x, ...)
}
\arguments{
\item{x}{mars object x}

\item{...}{other arguments like line size and color}
}
\value{
returns a plot of fitted values
}
\description{
plots the fitted basis that depends on one explanatory variable or two explanatory variables.
}
\examples{
m<- mars(wage~age+education,data=ISLR::Wage)
plot(m,col="red")
class(m)<-"lm";plot(m) #standard diagnostic plots
}
\seealso{
Other methods: 
\code{\link{print.mars}()}
}
\concept{methods}
