% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot.mars}
\usage{
\method{plot}{mars}(object)
}
\arguments{
\item{object}{an \code{mars} object from the result from mars() function}
}
\description{
generates four different residual plots
}
\details{
The Cumulative Distribution plot shows the empirical cumulative distribution function (CDF) of the data. The empirical CDF is the ratio of values ​​less than or equal to X. It is an increasing step function that exhibits a vertical jump of 1/N for each value of X equal to the observed value.
The Residuals vs. Fit plot is a scatterplot of residuals on the y-axis and fitted values ​​(estimated responses) on the x-axis. The chart detects nonlinearities, uneven error variances, and outliers.
Normal Q-Q plots are used to assess how well the distribution of a data set fits a normal (Gaussian) standard distribution with mean  0 and  standard deviation  1.
The Scale-Location plot plots a function of the residuals that reflects the  variability of the error about the mean. I'm looking for a point cloud that looks random. An increasing relationship means that as the mean increases the variance increases  and vice versa.
}
\examples{
test <- mars(y~.,data=data1[1:100,])

plot(test)
}
