% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mars}
\alias{plot.mars}
\title{plot.mars}
\usage{
S3 method for class 'print.plot'
}
\arguments{
\item{x}{a mars obeject}

\item{...}{additional arguments to pass to plot()}
}
\value{
Plots of the fitted model of normal distribution and showing the linear relationships.
}
\description{
Plots the fitted basis functions that depend on one explanatory variable (main effects) or two explanatory variables (two-way interactions). Use 'predict.lm()' to see the residual plots
}
\details{
Plot a mars object. By default the plot shows residuals versus fitted values.
}
\examples{
mm <- mars(y~x1+x2, data = marstestdata, mars.control(Mmax=4))
class(mm) <- 'lm'; plot(mm)
mars::plot.mars(marstestdata)
}
\seealso{
\link{mars.control} for constructing the control object.

\link{predict.mars} for predictions

\link{summary.mars}/\link{print.mars} for summarizing

Other methods: 
\code{\link{mars}()},
\code{\link{predict.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
