% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot.mars}
\usage{
\method{plot}{mars}(object)
}
\arguments{
\item{object}{an \code{mars} object from the result from mars() function}
}
\description{
generates four different residual plots
}
\details{
A Cumulative Distribution plot shows the empirical cumulative distribution function (CDF) of the
data. The empirical CDF is the proportion of values less than or equal to X. It is an increasing step
function that has a vertical jump of 1/N at each value of X equal to an observed value.

A Residuals vs Fitted plot is a scatter plot of residuals on the y-axis and fitted values (estimated
responses) on the x-axis. The plot detects non-linearity, unequal error variances, and outliers.

A Normal Q-Q plot is used to evaluate how well the distribution of a dataset matches a standard
normal (Gaussian) distribution, where the mean is 0 and the standard deviation is 1.

A Scale-Location plot plots a function of residuals that reflect the errors' variability relative to the
mean, we will be looking for a random-looking cloud of point, increasing relationship means variance
increases as mean increases, and vice versa.
}
\examples{
test <- mars(y~.,data=data1[1:100,])

plot(test)
}
