% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict method for mars objects}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{a mars object to use for prediction, usually created by mars()}

\item{newdata}{optional data to use for prediction}
}
\description{
Predict method for mars objects
}
\seealso{
Other methods: 
\code{\link{anova.mars}()},
\code{\link{plot.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
