% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict method for mars objects}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{a mars object to use for prediction, usually created by mars()}

\item{newdata}{ptional data to use for prediction}
}
\value{
a vector of prediction values
}
\description{
Predict method for mars objects
}
\examples{
## Predict values for iris data
m <-mars(Sepal.Length~Sepal.Width+Petal.Length+Petal.Width,data=iris,control=mars.control())
predict(m)
}
\seealso{
Other methods: 
\code{\link{anova.mars}()},
\code{\link{plot.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
