% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predicted values based on mars object.}
\usage{
\method{predict}{mars}(object, newdata, ...)
}
\arguments{
\item{object}{a \code{mars} object}

\item{newdata}{An optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{further arguments}
}
\value{
predicted values of the response variable
}
\description{
Predicted values based on mars object.
}
\examples{
mar <- mars(y~x1+x2+x3, data=mars::marstestdata)
predict(object=mar, newdata=mars::marstestdata)
}
\seealso{
Other methods: 
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
