% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mars}
\alias{predict.mars}
\title{Prediction Function for MARS Model}
\usage{
\method{predict}{mars}(object, newdata, ...)
}
\arguments{
\item{object}{MARS model object (input mars object)}

\item{newdata}{data frame or matrix with new data}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
predicted values of fitted model
}
\description{
Predicted values of fitted mars model.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
predict.mars(fit.mars)
}
\seealso{
\code{\link[=mars]{mars()}}

\code{\link[=print.mars]{print.mars()}}

\code{\link[=summary.mars]{summary.mars()}}

\code{\link[=plot.mars]{plot.mars()}}

\code{\link[=anova.mars]{anova.mars()}}
}
