% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict method for  Multivariate Adaptive Regression Splines (MARS) fits}
\usage{
\method{predict}{mars}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class "mars".}

\item{newdata}{an optional data frame in which to look for variables with which to predict.
If omitted, the fitted values are used.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
produces a vector of predictions.
}
\description{
predicted values based on Multivariate Adaptive Regression Splines (MARS) objects.
}
\details{
predict.mars produces predicted values. It takes in as argument an object of class "mars", and use the
fit from this object to produce predictions on the test data provided in the newdata argument. If newdata is NULL,
prediction will be done on the training data (the one used to create object argument). There will be no need to create
a new B matrix as B matrix from object argument will be used. If the newdata argument is not NULL, then prediction will be
done on the new dataset, and new B matrix will be constructed for this particular dataset using Bfuncs list from fitted model.
}
\examples{
#if newdata is NULL
predict(object = testmars)

#if newdata is not NULL
predict(object = testmars, newdata = marstestdata)
}
\references{
Friedman, J.H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, Vol. 19, No. 1 (Mar., 1991), pp. 1-67
}
\seealso{
"Multivariate Adaptive Regression Splines (MARS)" to create an object of class "mars", plot.mars method to plot mars objects, print.mars method to print mars objects and summary.mars method to summarize mars objects.
}
\author{
Samir Arora and Gurpal Singh Tulli
}
