% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict Function}
\usage{
\method{predict}{mars}(object, newdata, ...)
}
\arguments{
\item{object}{a mars object}

\item{newdata}{optional new data for which predictions are required}

\item{...}{additional arguments to predict()--currently not used}
}
\value{
the corresponding matrix of basis functions
}
\description{
Predict Function
}
\examples{
mm<- mars(y~x1+x2,dat=mars::marstestdata)
predict(mm,newdata=data.frame(x1=rnorm(100),x2=rnorm(100)))
}
\seealso{
Other methods: 
\code{\link{plot.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
