% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict method for the MARS object}
\usage{
\method{predict}{mars}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{mars} object}

\item{newdata}{New data *\emph{optional}}

\item{...}{Additional arguments}
}
\value{
If \code{newdata} is missing, fitted values are returned;
otherwise, predicted values on the new data are returned
}
\description{
Provides predictions from the results of the MARS object's basis functions.
}
\examples{
mm <- mars(y~x1+x2, data=mars::marstestdata)
pred <- predict(mm, newdata=data.frame(x1=rnorm(100), x2=rnorm(100)))
}
\seealso{
Other method: 
\code{\link{mars}()},
\code{\link{plot.mars}()},
\code{\link{print.mars}()},
\code{\link{summary.mars}()}
}
\concept{method}
