% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Creates a prediction for Y given new data}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{an mars objects}

\item{newdata}{additional data with m rows}
}
\value{
A vector length m with predicted values
}
\description{
Creates a prediction for Y given new data
}
\examples{
# see facebook data ex. in test file
predict.mars(object, newdata)
}
