% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Creates a prediction for Y given new data}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{an mars objects}

\item{newdata}{additional data with m rows}
}
\value{
A vector length m with predicted values
}
\description{
Creates a prediction for Y given new data
}
\examples{
#facebook data
df3 <- read.delim("./dataset_Facebook.csv",sep=";",header=TRUE)
df3 <- subset(df3, select = c(like, comment, share))
df3 <- head(df3,400)
newdata <- tail(df3,100)
names(newdata) <- names(df3)
object <- mars(like~., df3)
predict.mars(object, newdata)
}
