% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict method for MARS}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{An object of class mars}

\item{newdata}{A data frame to predict with}
}
\description{
Predict with an mars model for new data, returns the predicted basis function.
}
\examples{
mc = mars.control()
mm = mars(y ~ .,data=marstestdata,control=mc)
}
\author{
Ajay Behal Lei Liu Linden Gueck
}
