% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict Method for mars object}
\usage{
\method{predict}{mars}(object, newdata, ...)
}
\arguments{
\item{object}{a mars object}

\item{newdata}{optional specification of new data on which prediction could be derived}

\item{...}{other arguments}
}
\value{
a vector of prediction values
}
\description{
Predicted Values
}
\examples{
m<- mars(wage~age+education,data=ISLR::Wage)
predict(m)

}
