% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mars}
\alias{predict.mars}
\title{predict.mars(): Predict values from a fitted mars object}
\usage{
S3 method for class 'print.predict'
}
\arguments{
\item{object}{a mars object}

\item{newdata}{optional new data for which predictions are required}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
a vector of predictions
}
\description{
Predicted values based on multivariate adaptive regression spline models(mars).
}
\details{
It is based on \code{predict.lm()}It is based on \code{predict.lm()},predictions from the fitted model in testmars on “new data” in marstestdata.This is the same data used to fit the model, but by passing it in as newdata we test that part of the code.Compare the output of this call to the expected output testpredict.
}
\examples{
mm <- mars(y~x1+x2, dat = mars::marstestdata)
predict(mm, newdata=data.frame(x1=rnorm(100), x2= rnorm(100)))
}
\seealso{
\link{mars.control} for constructing the control object.

\link{plot.mars} for plotting the results

\link{summary.mars}/\link{print.mars} for summarizing

Other methods: 
\code{\link{mars}()},
\code{\link{plot.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
