% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Print values for a mars Object}
\usage{
\method{print}{mars}(mars)
}
\arguments{
\item{mars}{a mars object, usually created by mars}
}
\description{
Print values for a mars Object
}
