% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Prints out the function call and coefficients of a \code{mars} object}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{a \code{mars} object}

\item{...}{further arguments}
}
\description{
Prints out the function call and coefficients of a \code{mars} object
}
\examples{
mar <- mars(y~x1+x2+x3, data=mars::marstestdata, control=mars.control(Mmax=6))
print(mar)
}
\seealso{
Other methods: 
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
