% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mars}
\alias{print.mars}
\title{Print method for MARS Model}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{MARS model object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print coefficients of the fitted mars model.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
print.mars(fit.mars)

}
