% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mars}
\alias{print.mars}
\title{Print method for MARS Model}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{MARS model object (input mars object)}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print coefficients of the fitted mars model.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
print.mars(fit.mars)

#example 2
# create your own dataset
n <- 1000
x1 <- rnorm(n)
x2 <- rnorm(n)
x3 <- rnorm(n)
y <- 2*x1 + 3*x2 + 4*x3 + rnorm(n)
df <- data.frame(x1=x1,x2=x2,x3=x3,y=y)

# fit a MARS model
fit <- mars(y ~ x1 + x2 + x3, data=df)
#show results
print.mars(fit)

}
\seealso{
\code{\link[=mars]{mars()}}

\code{\link[=summary.mars]{summary.mars()}}

\code{\link[=predict.mars]{predict.mars()}}

\code{\link[=plot.mars]{plot.mars()}}

\code{\link[=anova.mars]{anova.mars()}}
}
