% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mars}
\alias{print.mars}
\title{Print method for Multivariate Adaptive Regression Splines (MARS) fits}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{an x of class "mars".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
There is no return value.
}
\description{
print method for class "mars"
}
\details{
print.mars method prints two important pieces of information about the x of class "mars".
First is the call made by user to mars function, second is the coefficients of the fitted model in the x argument.
}
\examples{
print(testmars)
}
\references{
Friedman, J.H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, Vol. 19, No. 1 (Mar., 1991), pp. 1-67
}
\seealso{
"Multivariate Adaptive Regression Splines (MARS)" to create an x of class "mars", plot.mars method to plot mars xs, predict.mars method to prediction,
and summary.mars method to summarize mars xs.
}
\author{
Samir Arora and Gurpal Singh Tulli
}
