% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Print a Mars Object}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{a \code{mars} object}

\item{...}{additional arguments to print.mars, currently unused}
}
\description{
Print a Mars Object
}
\examples{
mm <- mars(y~x1+x2,data=marstestdata)
print(mm)
}
\seealso{
Other methods: 
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{summary.mars}()}
}
\concept{methods}
