% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mars}
\alias{print.mars}
\title{print.mars}
\usage{
\method{print}{mars}(marsobject, ...)
}
\value{
values of the coefficients for MARS
}
\description{
PRINT METHOD MARS OBJECT
}
\details{
Prints intercept and coefficient of mars object
}
\examples{
mm <- mars(y~x1+x2,data=marstestdata)
print(mm)
}
\author{
Tauseef Kashtwari, Promit Chowdhury, Ibraheem Azad
}
