% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Print method for the MARS object}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{A \code{mars} object}

\item{...}{Additional arguments}
}
\value{
Function call and coefficients of the MARS object
}
\description{
Prints the function call of MARS object, then each coefficient of selected variables.
}
\examples{
mm <- mars(y~., data=mars::marstestdata)
print(mm)
}
\seealso{
Other method: 
\code{\link{mars}()},
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{summary.mars}()}
}
\concept{method}
