% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{print.mars}
\alias{print.mars}
\title{Prints the mars object}
\usage{
\method{print}{mars}(mars)
}
\arguments{
\item{mars}{an mars objects}

\item{data}{a data frame containing the data}
}
\value{
void
}
\description{
Prints the mars object
}
\examples{
print(mars)
}
