% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.mars}
\alias{print.mars}
\title{Prints out a mars object}
\usage{
\method{print}{mars}(marsobject, ...)
}
\value{
values of the coefficients
}
\description{
print mars object
}
\details{
Prints intercept and coefficient of mars object
}
\examples{
mm <- mars(y~x1+x2,data=marstestdata)
print(mm)
}
\author{
Ajay Behal Lei Liu Linden Gueck
}
