% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Print method for mars object}
\usage{
\method{print}{mars}(object, ...)
}
\arguments{
\item{object}{a mars object}

\item{...}{other additional arguments to print.mars}
}
\value{
returns the call to mars and  estimated basis function coefficients.
}
\description{
Print method for mars object
}
\examples{
m<- mars(wage~age+education,data=ISLR::Wage)
print(m)
}
\seealso{
Other methods: 
\code{\link{plot.mars}()}
}
\concept{methods}
