% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Print mars}
\usage{
\method{print}{mars}(object)
}
\arguments{
\item{object}{an \code{mars} object from the result from mars() function}
}
\description{
Print some information for a \code{mars} object, include calls formula and coefficients
}
\details{
First calls the mars() function and the regression model formula, then provides the coefficients for each independent variable.
}
\examples{
test <- mars(y~.,data=data1[1:100,])
print(test)

}
