% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\alias{profile.MARSassess}
\alias{plot.MARSprof}
\title{Profile parameters of MARS model}
\usage{
\method{profile}{MARSassess}(fitted, p1, v1, p2, v2, ...)

\method{plot}{MARSprof}(x, component = "objective", rel = TRUE, xlab, ylab, main, ...)
}
\arguments{
\item{fitted}{\linkS4class{MARSassess} object returned by \code{\link[=fit_MARS]{fit_MARS()}}}

\item{p1}{Character string that represents the first parameter to be profiled,
including the parameter name and index. See "Parameters" section of \code{\link[=make_parameters]{make_parameters()}}}

\item{v1}{Vector of values corresponding to \code{p1}}

\item{p2}{Character string that represents the optional second parameter to be profiled}

\item{v2}{Vector of values corresponding to \code{p2}}

\item{...}{Other argument to base graphics}

\item{x}{Output from \code{\link[=profile.MARSassess]{profile.MARSassess()}}}

\item{component}{Character for the column in \code{x} to be plotted}

\item{rel}{Logical, whether the relative change in \code{component} is plotted (TRUE) or the raw values (FALSE)}

\item{xlab}{Optional character for the x-axis label}

\item{ylab}{Optional character for the y-axis label}

\item{main}{Optional character for the plot title}
}
\value{
The profile generic returns a data frame of the likelihood values that correspond to
fixed values of \code{p1} and \code{p2}.

The accompanying plot function returns a line plot for a 1-dimensional profile or a contour plot for a two
dimensional profile.
}
\description{
Evaluate change in objective function and likelihood components for up to 2 parameters.
}
