% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_mean.R
\name{running_mean-file}
\alias{running_mean-file}
\title{Running mean functions}
\description{
running mean functions (also for matrix) internally uses the smear functions for speed/space and then divides by the amount of position summed over.
}
\author{
Mark Heron
}
