% Generated by roxygen2 (4.0.1): do not edit by hand
\name{running_mean_ff_matrix}
\alias{running_mean_ff_matrix}
\title{running mean on ff_matrix}
\usage{
running_mean_ff_matrix(mat, smooth, over = "rows")
}
\description{
running mean on ff_matrix
}
\examples{
running_mean_ff_matrix(as.ff(matrix(1:16,4,4)), 3, over="rows")
}
\seealso{
running_mean_matirx
}

