% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_mean.R
\name{running_mean_ff_matrix}
\alias{running_mean_ff_matrix}
\title{running mean on ff_matrix}
\usage{
running_mean_ff_matrix(mat, smooth, over = "rows")
}
\arguments{
\item{mat}{(ff_matrix) matrix to smooth}

\item{smooth}{(int) the smothing window size}

\item{over}{(character) if the matrix should be smoothed over (i.e. along) the rows (first dimension) or columns (second dimension)}
}
\description{
running mean on ff_matrix
}
\examples{
library(ff)
running_mean_ff_matrix(as.ff(matrix(1:16,4,4)), 3, over="rows")

}
\seealso{
running_mean_matirx
}

