% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smear.R
\docType{methods}
\name{smear}
\alias{smear}
\alias{smear,numeric,matrix-method}
\alias{smear,numeric-method}
\alias{smear,matrix-method}
\title{smear}
\usage{
smear(x, from = 0, to = 1)

\S4method{smear}{numeric}(x, from = 0, to = 1)

\S4method{smear}{matrix}(x, from = 0, to = 1)
}
\arguments{
\item{x}{numeric data to smear}

\item{from}{where to start}

\item{to}{where to go}
}
\value{
smeared x
}
\description{
Computes a smeared version of the data, basically a running sum.
}
\details{
Adds the data shifted onto itself several times, using a tree/recursive approach so the time complexity is only \strong{log(to-from)}.
The start and end will end up with smaller values on average, since the ends are extended with zeros.

\subsection{if x is numeric vector}{
The vector is shifted as expected.
}

\subsection{if x is numeric matrix}{
The matrix is shifted along the first dimension i.e. the smearing happens along the rows.
}
}
