% Generated by roxygen2 (4.0.1): do not edit by hand
\name{smear_ff_simple}
\alias{smear_ff_simple}
\title{smear_ff_simple}
\usage{
smear_ff_simple(x, from = 0, to = 1)
}
\arguments{
\item{x}{ff vector to smear}

\item{from}{where to start}

\item{to}{where to go}
}
\value{
smeared ff vector
}
\description{
Computes a smeared vector, basically a running_sum.
}
\details{
Add's the vector on its self several times, so good for short smears of long vectors.
}
\author{
Mark Heron
}

