% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Produce result summaries of a mars object}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{a \code{mars} object for which a summary is desired.}

\item{...}{further arguments}
}
\description{
Prints Call, Five-number summary, summary of hinge functions for each basis function,
and the coefficients of each basis function of a \code{mars} object
}
\examples{
mar <- mars(y~x1+x2+x3, data=dataset)
summary(mar)
}
\seealso{
Other methods: 
\code{\link{predict.mars}()},
\code{\link{print.mars}()}
}
\concept{methods}
