% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary Function for MARS Model}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{a mars object}

\item{...}{further arguments passed to or from other methods}
}
\value{
The function return the summary of the input MARS object as well as
a generic summary function with the MARS object
}
\description{
Summary Function for MARS Model
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
summary.mars(fit.mars)
}
