% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary Function for MARS Model}
\usage{
\method{summary}{mars}(object,...)
}
\arguments{
\item{object}{MARS model object (input mars object)}

\item{...}{further arguments passed to or from other methods}
}
\value{
The function return the summary of the input MARS object as well as
a generic summary function with the MARS object
}
\description{
summary method for class 'mars'.
}
\examples{
fit.mars <- mars(wage ~ age + education, data=ISLR::Wage, control = mars.control(Mmax=10))
summary.mars(fit.mars)
}
\seealso{
\code{\link[=mars]{mars()}}

\code{\link[=print.mars]{print.mars()}}

\code{\link[=predict.mars]{predict.mars()}}

\code{\link[=plot.mars]{plot.mars()}}

\code{\link[=anova.mars]{anova.mars()}}
}
