% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mars}
\alias{summary.mars}
\title{Summarizing Multivariate Adaptive Regression Splines (MARS) fits}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{an obejct of class "mars".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
There is no return value for this method.
}
\description{
summary method for class "mars".
}
\details{
summary.mars method presents important parts of an object of class "mars" in a very
structured way. It presents the call made by user to mars function, coefficients of the fitted model,
basis functions and component of each and every basis function presented in the final model.
}
\examples{
summary(testmars)
}
\references{
Friedman, J.H. (1991). Multivariate Adaptive Regression Splines. The Annals of Statistics, Vol. 19, No. 1 (Mar., 1991), pp. 1-67
}
\seealso{
"Multivariate Adaptive Regression Splines (MARS)" to create an object of class "mars", plot.mars method to plot mars objects, predict.mars method for
prediction, and print.mars method to print mars objects.
}
\author{
Samir Arora and Gurpal Singh Tulli
}
