% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary of Mars Object}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{an object of class "mars", usually a result of a call to \link{mars}}

\item{...}{additional arguments to print.mars, currently unused}
}
\value{
a summary of the mars object
}
\description{
summary method for class "mars". Prints a summary of hinges that make up each basis function in the optimal model, along with the coefficient of said basis functions
}
\examples{
mm<-mars(y ~.,data=mars::marstestdata)
summary(mm)

}
\seealso{
Other methods: 
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{print.mars}()}
}
\author{
Gareth Bennett, Heewon Oh, Jusung Lee
}
\concept{methods}
