% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary for MARS object}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{A \code{mars} object}

\item{...}{Additional arguments}
}
\value{
A summary statistic of the MARS object including:
\itemize{
\item Function call
\item Quarterly summary of residual values
\item Coefficients of variables
\item Residual standard error
}
}
\description{
Prints a summary statistic of the MARS object, from the data and fitted model.
}
\examples{
mm <- mars(y~., data=mars::marstestdata)
summary(mm)
}
\seealso{
Other method: 
\code{\link{mars}()},
\code{\link{plot.mars}()},
\code{\link{predict.mars}()},
\code{\link{print.mars}()}
}
\concept{method}
