% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mars}
\alias{summary.mars}
\title{summary.mars}
\usage{
\method{summary}{mars}(object, digits)
}
\arguments{
\item{object}{of class mars, which is obtained from calling mars()}

\item{digits}{the number of significant digits i.e. SIG FIG}
}
\description{
prints a summary of the mars object with the function call. Also prints the summary of the hinges of each basis function.
}
\examples{
mm<-mars(y ~.,data=mars::marstestdata)
summary(mm)
}
\author{
Ajay Behal Lei Liu Linden Gueck
}
