% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{summary.mars}
\usage{
\method{summary}{mars}(object, correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
summary.mars
}
