% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary method for mars object}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{a mars object}

\item{...}{other optional arguments to be passed to summary.mars}
}
\value{
the call to mars, summary of residuals of the fitted values, estimated basis function coefficients and residual standard error
}
\description{
Summary method for mars object
}
\examples{
m<- mars(wage~age+education,data=ISLR::Wage)
summary(m)
}
