% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mars}
\alias{summary.mars}
\title{: summary.mars}
\usage{
S3 method for class 'summary.mars'
}
\arguments{
\item{object}{a mars object output by the mars()}

\item{...}{additional arguments}
}
\value{
a summary of the mars object
}
\description{
: This function summary method for mars objects.
}
\details{
Summary method for the mars objects and function in the final model and the coefficients of each basis function.
}
\examples{
summary.mars(marstestdata)
}
\seealso{
\link{mars.control} for constructing the control object.

\link{plot.mars} for plotting the results.

\link{predict.mars} for predictions

\link{print.mars} for summarizing results.

Other methods: 
\code{\link{mars}()},
\code{\link{plot.mars}()},
\code{\link{predict.mars}()}
}
\concept{methods}
