% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Summary Mars}
\usage{
\method{summary}{mars}(object)
}
\arguments{
\item{object}{an \code{mars} object from the result from mars() function}
}
\description{
provides a summary table of the estimated coefficients of all the independent variables with respect to the dependent variable.
}
\details{
For each independent variable, a sign (represents where the split direction is left: -1 or right: +1) and a knot for each split are also provided.
}
\examples{
test <- mars(y~.,data=data1[1:100,])

summary(test)
}
