% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_for_inclined_surface.R
\name{x_for_inclined_surface}
\alias{x_for_inclined_surface}
\title{Title}
\usage{
x_for_inclined_surface(phi, beta, gamma_c)
}
\arguments{
\item{phi}{}

\item{beta}{}

\item{gamma_c}{}
}
\value{

}
\description{
Title
}
