% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z.R
\name{Z}
\alias{Z}
\title{Equation 6: Zenith angle of the incident solar radiation [deg].}
\usage{
Z(Ls, T_s, phi, nfft)
}
\arguments{
\item{Ls}{Areocentric longitude [deg].}

\item{T_s}{Solar time [h]}

\item{phi}{Planetary latitude [deg].}

\item{nfft}{Net flux function type.}
}
\value{

}
\description{
Source: Equation 6 in Appelbaum, Joseph & Flood, Dennis. (1990).
}
